Crate lurk

Re-exports

pub use symbol::Symbol;

pub use uint::UInt;

pub use z_data::z_cont;

pub use z_data::z_expr;

pub use z_data::z_ptr;

pub use z_data::z_store;

Modules

circuit

cli

config

Global config for Lurk Includes settings for cache locations, public parameters, and parallelism.

cont

coprocessor

error

eval

expr

field

The finite field used in the language.

hash

hash_witness

lem

Lurk Evaluation Model (LEM)

package

parser

proof

This module offers a connection the the backend proving engine of Lurk. Abstracted behind the Prover and Verifier traits, the proving engine has two instantiations:

ptr

public_parameters

repl

state

This module implements an abstraction for the Lurk state, which changes as Lurk code is evaluated

store

symbol

syntax

tag

uint

writer

z_data

ZData is a lightweight binary data serialization format.

Macros

block

char

ctrl

func

keyword

list

lit

lurk_sym_ptr

num

op

str

sym

symbol

tag

uint

var

vars

Enums

Num

Finite field element type for Lurk. Has different internal representations to optimize evaluation.

Constants

TEST_SEED