Trait lurk::circuit::gadgets::circom::CircomGadget

pub trait CircomGadget<F: LurkField>: Send + Sync + Clone {
    // Required methods
    fn name(&self) -> &str;
    fn into_circom_input(
        self,
        input: &[AllocatedPtr<F>]
    ) -> Vec<(String, Vec<F>)>;
    fn simple_evaluate(&self, s: &Store<F>, args: &[Ptr<F>]) -> Ptr<F>;
}

An interface to declare a new type of Circom gadget. It requires 3 things:

The use defined name of the gadget. This must be an existing name loaded into the file system via the CLI (with lurk coprocessor --name <NAME> <CIRCOM_FOLDER>)

A defined way to take a list of Lurk input pointers and turn them into a Circom input. We do not enforce the shapes of either the Lurk end or the Circom end, so users should take care to define what shape they expect.

A defined way Lurk should evaluate what this gadget does. This is then the implementation used in the Coprocessor trait.

Required Methods

fn name(&self) -> &str

fn into_circom_input(self, input: &[AllocatedPtr<F>]) -> Vec<(String, Vec<F>)>

fn simple_evaluate(&self, s: &Store<F>, args: &[Ptr<F>]) -> Ptr<F>