Function lurk::config::lurk_config

pub fn lurk_config(
    file: Option<&Utf8PathBuf>,
    settings: Option<&HashMap<&str, String>>
) -> &'static Settings

Gets the LURK_CONFIG settings. If uninitialized, sets the global variable in the following order (greatest to least precedence):

settings map if provided, e.g. with key (“public_params_dir”, “$HOME/lurk-rs/public_params”)

Env var per setting, e.g. LURK_PUBLIC_PARAMS_DIR

Config file, which also has a configurable location (see lurk_config_file()), and has the following syntax for e.g. TOML:

public_params_dir = "/path/to/params"

Other file formats are supported by the config crate, but only TOML is tested

Default values, e.g. $HOME/.lurk/public_params