Module lurk::coprocessor

Modules

circom

trie

The trie module implements a Trie with the following properties: Big-endian bits of the (field element) key are taken N at a time, where the underlying hash has arity 2^N. This forms a fixed-length path whose final element is either zero or some non-zero payload. If the payload is assumed to be a commitment, then zero can be used as a hash whose preimage is undiscoverable, and hence empty. To optimize creation of an empty tree, we precompute the empty subtree at each row rather than trying to actually construct the tree with a vast number of redundant hashes.

Traits

CoCircuit

CoCircuit is a trait that represents a generalized interface for coprocessors. Coprocessors augment the Lurk circuit and evaluation with additional built-in functionality. This trait generalizes over functionality needed in the circuit, the sibling Coprocessor trait, generalizes over functionality needed in the evaluator.

Coprocessor

Coprocessor is a trait that represents a generalized interface for coprocessors. Coprocessors augment the Lurk circuit and evaluation with additional built-in functionality. This trait generalizes over functionality needed in the evaluator, the sibling CoCircuit trait, generalizes over functionality needed in the circuit.