Module lurk::coprocessor::gadgets

Helper gadgets for synthesis

Functions

a_ptr_as_z_ptr

car_cdr

Deconstructs data with car_cdr semantics.

chain_car_cdr

Chains car_cdr calls n times, returning the accumulated cars, the final cdr and the (explored) actual length (<= n) of the cons-like data. For example, calling chain_car_cdr on “ab” with n = 4 should return the full actual length 2 of such string. But if n is set to 1, it will return 1 as the (explored) length.

construct_cons

Constructs a Cons pointer

construct_list

Constructs a cons-list with the provided elts. The terminating value defaults to nil when last is None

construct_tuple2

Constructs an AllocatedPtr compound by two others

construct_tuple3

Constructs an AllocatedPtr compound by three others

construct_tuple4

Constructs an AllocatedPtr compound by four others

deconstruct_tuple2

Deconstructs tuple, assumed to be a composition of two others.

deconstruct_tuple3

Deconstructs tuple, assumed to be a composition of three others.

deconstruct_tuple4

Deconstructs tuple, assumed to be a composition of four others.