Function lurk::coprocessor::gadgets::chain_car_cdr

pub fn chain_car_cdr<F: LurkField, CS: ConstraintSystem<F>>(
    cs: &mut CS,
    g: &GlobalAllocator<F>,
    store: &Store<F>,
    not_dummy: &Boolean,
    data: &AllocatedPtr<F>,
    n: usize
) -> Result<(Vec<AllocatedPtr<F>>, AllocatedPtr<F>, AllocatedNum<F>), SynthesisError>

Chains car_cdr calls n times, returning the accumulated cars, the final cdr and the (explored) actual length (<= n) of the cons-like data. For example, calling chain_car_cdr on “ab” with n = 4 should return the full actual length 2 of such string. But if n is set to 1, it will return 1 as the (explored) length.

It can be used to deconstruct cons-lists into their elements or strings into their characters.

Panics

Panics if any of the reached elements can’t be deconstructed with car_cdr

use lurk::{
    coprocessor::gadgets::{a_ptr_as_z_ptr, chain_car_cdr},
};

let store = Store::<Fq>::default();
let nil = store.intern_nil();
let z_nil = store.hash_ptr(&nil);
let empty_str = store.intern_string("");
let z_empty_str = store.hash_ptr(&empty_str);
let not_dummy = Boolean::Constant(true);

let ab = store.intern_string("ab");
let z_ab = store.hash_ptr(&ab);
let a = Ptr::char('a');
let b = Ptr::char('b');
let z_a = store.hash_ptr(&a);
let z_b = store.hash_ptr(&b);
let a_ab = AllocatedPtr::alloc_infallible(&mut cs.namespace(|| "ab"), || z_ab);
let (cars, cdr, length) = chain_car_cdr(
    &mut cs.namespace(|| "chain_car_cdr on ab"),
    &g,
    &store,
    &not_dummy,
    &a_ab,
    4,
)
.unwrap();
assert_eq!(cars.len(), 4);
assert_eq!(a_ptr_as_z_ptr(&cars[0]), Some(z_a));
assert_eq!(a_ptr_as_z_ptr(&cars[1]), Some(z_b));
assert_eq!(a_ptr_as_z_ptr(&cars[2]), Some(z_nil));
assert_eq!(a_ptr_as_z_ptr(&cars[3]), Some(z_nil));
assert_eq!(a_ptr_as_z_ptr(&cdr), Some(z_empty_str));
assert_eq!(length.get_value(), Some(Fq::from_u64(2)));