Trait lurk::coprocessor::CoCircuit

pub trait CoCircuit<F: LurkField>: Send + Sync + Clone {
    // Provided methods
    fn arity(&self) -> usize { ... }
    fn synthesize_alpha<CS: ConstraintSystem<F>>(
        &self,
        _cs: &mut CS,
        _g: &GlobalAllocations<F>,
        _store: &AlphaStore<F>,
        _input_exprs: &[AllocatedPtr<F>],
        _input_env: &AllocatedPtr<F>,
        _input_cont: &AllocatedContPtr<F>,
        _not_dummy: &Boolean
    ) -> Result<(AllocatedPtr<F>, AllocatedPtr<F>, AllocatedContPtr<F>), SynthesisError> { ... }
    fn synthesize_internal<CS: ConstraintSystem<F>>(
        &self,
        cs: &mut CS,
        g: &GlobalAllocator<F>,
        s: &Store<F>,
        not_dummy: &Boolean,
        ptrs: &[AllocatedPtr<F>]
    ) -> Result<Vec<AllocatedPtr<F>>, SynthesisError> { ... }
    fn synthesize<CS: ConstraintSystem<F>>(
        &self,
        cs: &mut CS,
        g: &GlobalAllocator<F>,
        s: &Store<F>,
        not_dummy: &Boolean,
        args: &[AllocatedPtr<F>],
        env: &AllocatedPtr<F>,
        cont: &AllocatedPtr<F>
    ) -> Result<Vec<AllocatedPtr<F>>, SynthesisError> { ... }
    fn synthesize_simple<CS: ConstraintSystem<F>>(
        &self,
        _cs: &mut CS,
        _g: &GlobalAllocator<F>,
        _s: &Store<F>,
        _not_dummy: &Boolean,
        _args: &[AllocatedPtr<F>]
    ) -> Result<AllocatedPtr<F>, SynthesisError> { ... }
}

CoCircuit is a trait that represents a generalized interface for coprocessors. Coprocessors augment the Lurk circuit and evaluation with additional built-in functionality. This trait generalizes over functionality needed in the circuit, the sibling Coprocessor trait, generalizes over functionality needed in the evaluator.

The trait is implemented by concrete coprocessor types, such as DumbCoprocessor.

Provided Methods

fn arity(&self) -> usize

fn synthesize_alpha<CS: ConstraintSystem<F>>( &self, _cs: &mut CS, _g: &GlobalAllocations<F>, _store: &AlphaStore<F>, _input_exprs: &[AllocatedPtr<F>], _input_env: &AllocatedPtr<F>, _input_cont: &AllocatedContPtr<F>, _not_dummy: &Boolean ) -> Result<(AllocatedPtr<F>, AllocatedPtr<F>, AllocatedContPtr<F>), SynthesisError>

fn synthesize_internal<CS: ConstraintSystem<F>>( &self, cs: &mut CS, g: &GlobalAllocator<F>, s: &Store<F>, not_dummy: &Boolean, ptrs: &[AllocatedPtr<F>] ) -> Result<Vec<AllocatedPtr<F>>, SynthesisError>

fn synthesize<CS: ConstraintSystem<F>>( &self, cs: &mut CS, g: &GlobalAllocator<F>, s: &Store<F>, not_dummy: &Boolean, args: &[AllocatedPtr<F>], env: &AllocatedPtr<F>, cont: &AllocatedPtr<F> ) -> Result<Vec<AllocatedPtr<F>>, SynthesisError>

fn synthesize_simple<CS: ConstraintSystem<F>>( &self, _cs: &mut CS, _g: &GlobalAllocator<F>, _s: &Store<F>, _not_dummy: &Boolean, _args: &[AllocatedPtr<F>] ) -> Result<AllocatedPtr<F>, SynthesisError>

Object Safety

This trait is not object safe.