Trait lurk::coprocessor::Coprocessor

pub trait Coprocessor<F: LurkField>: Clone + Debug + Sync + Send + CoCircuit<F> {
    // Required methods
    fn eval_arity(&self) -> usize;
    fn simple_evaluate(&self, s: &mut Store<F>, args: &[Ptr<F>]) -> Ptr<F>;

    // Provided methods
    fn evaluate(
        &self,
        s: &mut Store<F>,
        args: Ptr<F>,
        env: Ptr<F>,
        cont: ContPtr<F>
    ) -> IO<F> { ... }
    fn has_circuit(&self) -> bool { ... }
    fn synthesize_step_circuit<CS: ConstraintSystem<F>>(
        &self,
        cs: &mut CS,
        store: &Store<F>,
        g: &GlobalAllocations<F>,
        input_expr: &AllocatedPtr<F>,
        input_env: &AllocatedPtr<F>,
        input_cont: &AllocatedContPtr<F>
    ) -> Result<(AllocatedPtr<F>, AllocatedPtr<F>, AllocatedContPtr<F>), SynthesisError> { ... }
}

Coprocessor is a trait that represents a generalized interface for coprocessors. Coprocessors augment the Lurk circuit and evaluation with additional built-in functionality. This trait generalizes over functionality needed in the evaluator, the sibling CoCircuit trait, generalizes over functionality needed in the circuit.

The trait is implemented by concrete coprocessor types, such as DummyCoprocessor.

Pattern

We use a type class (trait) pattern to provide extensibility for the coprocessor implementations. The pattern involves:

A trait crate::coprocessor::Coprocessor, which defines the methods and behavior for all coprocessors.

An enum such as crate::eval::lang::Coproc, which “closes” the hierarchy of possible coprocessor implementations we want to instantiate at a particular point in the code.

Required Methods

fn eval_arity(&self) -> usize

fn simple_evaluate(&self, s: &mut Store<F>, args: &[Ptr<F>]) -> Ptr<F>

As with all evaluation, the value returned from simple_evaluate must be fully evaluated.

Provided Methods

fn evaluate( &self, s: &mut Store<F>, args: Ptr<F>, env: Ptr<F>, cont: ContPtr<F> ) -> IO<F>

fn has_circuit(&self) -> bool

Returns true if this Coprocessor actually implements a circuit.

fn synthesize_step_circuit<CS: ConstraintSystem<F>>( &self, cs: &mut CS, store: &Store<F>, g: &GlobalAllocations<F>, input_expr: &AllocatedPtr<F>, input_env: &AllocatedPtr<F>, input_cont: &AllocatedContPtr<F> ) -> Result<(AllocatedPtr<F>, AllocatedPtr<F>, AllocatedContPtr<F>), SynthesisError>