Type Alias lurk::coprocessor::trie::StandardTrie

pub type StandardTrie<'a, F> = Trie<'a, F, 8, 85>;

Aliased Type

struct StandardTrie<'a, F> { /* private fields */ }

Implementations

impl<'a, F: LurkField, const ARITY: usize, const HEIGHT: usize> Trie<'a, F, ARITY, HEIGHT>

pub fn root(&self) -> F

pub const fn leaves(&self) -> usize

How many leaves does this Trie have?

pub const fn row_size(&self, row: usize) -> usize

How many nodes does the rowth row have?

pub fn empty_root(&mut self) -> F

pub fn new(store: &'a Store<F>) -> Self

Creates a new Trie, saving preimage data in store. HEIGHT must be exactly that required to minimally store all elements of F.

pub fn new_with_capacity(store: &'a Store<F>, size: usize) -> Self

Creates a new Trie, saving preimage data in store. Height must be at least that required to store size elements.

pub fn lookup(&self, key: F) -> Result<Option<F>, Error<F>>

pub fn prove_lookup( &self, key: F ) -> Result<LookupProof<F, ARITY, HEIGHT>, Error<F>>

Returns a slice of preimages, corresponding to the path. Final preimage contains payloads.

pub fn insert(&mut self, key: F, value: F) -> Result<bool, Error<F>>

pub fn prove_insert( &mut self, key: F, value: F ) -> Result<(InsertProof<F, ARITY, HEIGHT>, bool), Error<F>>

pub fn synthesize_insert<CS: ConstraintSystem<F>>( &self, cs: &mut CS, s: &Store<F>, allocated_root: &AllocatedNum<F>, key: &AllocatedNum<F>, value: AllocatedNum<F> ) -> Result<AllocatedNum<F>, SynthesisError>