Module lurk::dual_channel

This module implements ChannelTerminal, meant to be used in pairs of its instances with crossed Senders and Receiver from mpsc::channel. This crossing is performed in pair_terminals. The idea is that one terminal can send/receive messages to/from the other.

Structs

ChannelTerminal

Holds a Sender and a Receiver which are not expected to be paired with each other

Functions

dummy_terminal

Creates a dummy ChannelTerminal that just sends messages to itself

pair_terminals

Creates a pair of ChannelTerminal with crossed senders and receivers such that one terminal can send/receive messages to/from the other