Trait lurk::field::LurkField

pub trait LurkField: PrimeField + PrimeFieldBits {
    const FIELD: LanguageField;

    // Provided methods
    fn to_bytes(self) -> Vec<u8>  { ... }
    fn from_bytes(bs: &[u8]) -> Option<Self> { ... }
    fn hex_digits(self) -> String { ... }
    fn trimmed_hex_digits(self) -> String { ... }
    fn to_u16(&self) -> Option<u16> { ... }
    fn to_u32(&self) -> Option<u32> { ... }
    fn to_char(&self) -> Option<char> { ... }
    fn to_u64(&self) -> Option<u64> { ... }
    fn to_u32_unchecked(&self) -> u32 { ... }
    fn to_u64_unchecked(&self) -> u64 { ... }
    fn from_u64(x: u64) -> Self { ... }
    fn from_u32(x: u32) -> Self { ... }
    fn from_u16(x: u16) -> Self { ... }
    fn from_char(x: char) -> Self { ... }
    fn most_negative() -> Self { ... }
    fn most_positive() -> Self { ... }
    fn is_negative(&self) -> bool { ... }
    fn from_expr_tag(tag: ExprTag) -> Self { ... }
    fn to_expr_tag(&self) -> Option<ExprTag> { ... }
    fn from_cont_tag(tag: ContTag) -> Self { ... }
    fn to_cont_tag(&self) -> Option<ContTag> { ... }
    fn from_op1(tag: Op1) -> Self { ... }
    fn to_op1(&self) -> Option<Op1> { ... }
    fn from_op2(tag: Op2) -> Self { ... }
    fn to_op2(&self) -> Option<Op2> { ... }
    fn get_field(&self) -> LanguageField { ... }

}

Required Associated Constants

const FIELD: LanguageField

Provided Methods

fn to_bytes(self) -> Vec<u8>

fn from_bytes(bs: &[u8]) -> Option<Self>

fn hex_digits(self) -> String

fn trimmed_hex_digits(self) -> String

fn to_u16(&self) -> Option<u16>

fn to_u32(&self) -> Option<u32>

fn to_char(&self) -> Option<char>

fn to_u64(&self) -> Option<u64>

fn to_u32_unchecked(&self) -> u32

fn to_u64_unchecked(&self) -> u64

fn from_u64(x: u64) -> Self

fn from_u32(x: u32) -> Self

fn from_u16(x: u16) -> Self

fn from_char(x: char) -> Self

fn most_negative() -> Self

fn most_positive() -> Self

0 - 1 is one minus the modulus, which must be even in a prime field. The result is the largest field element which is even when doubled. We define this to be the most positive field element.

fn is_negative(&self) -> bool

A field element is defined to be negative if it is odd after doubling.

fn from_expr_tag(tag: ExprTag) -> Self

fn to_expr_tag(&self) -> Option<ExprTag>

fn from_cont_tag(tag: ContTag) -> Self

fn to_cont_tag(&self) -> Option<ContTag>

fn from_op1(tag: Op1) -> Self

fn to_op1(&self) -> Option<Op1>

fn from_op2(tag: Op2) -> Self

fn to_op2(&self) -> Option<Op2>

fn get_field(&self) -> LanguageField