Module lurk::lem

Lurk Evaluation Model (LEM)

A LEM is a description of Lurk’s evaluation algorithm, encoded as data. In other words, it’s a meta-representation of Lurk’s step function.

The motivation behind LEM is the fact that hand-writing the circuit is a fragile process that hinders experimentation and safety. Thus we would like to bootstrap the circuit automatically, given a higher level description of the step function.

LEM also allows the Store API to be completely abstracted away from the responsibilities of LEM authors. Indeed, we want the implementation details of the Store not to be important at LEM definition time.

Data semantics

A LEM describes how to handle pointers with variables, which are basically named references. Instead of saying let foo ... in Rust, we use a Var("foo") in LEM.

The actual algorithm is encoded with a LEM operation (LEMOP). It’s worth noting that one of the LEM operators is in fact a vector of operators, which allows imperative/sequenced expressiveness.

Interpretation

Running a LEM is done via interpretation, which might be a bit slower than calling Rust functions directly. But it also has its advantages:

The logic to collect data during execution can be factored out from the definition of the step function. This process is needed in order to evidence the inputs for the circuit at proving time;

Actually, such logic to collect data is a natural consequence of the fact that we’re on a higher level of abstraction. Relevant data is not simply stored on rust variables that die after the function ends. On the contrary, all relevant data lives on data structures that are also a product of the interpreted LEM.

Constraining

This is the process of creating the circuit, which we want to be done automatically for whoever creates a LEM. Each LEMOP has to be precisely constrained in such a way that the resulting circuits accepts a witness iff it was generated by a valid interpretation of the LEM at play.

Static checks of correctness

Since a LEM is an algorithm encoded as data, we can perform static checks of correctness as some form of (automated) formal verification. Here are some (WIP) properties we want a LEM to have before we can adopt it as a proper Lurk step function:

Non-duplicated input labels: right at the start of interpretation, the input labels are bound to the actual pointers that represent the expression, environment and continuation. If some label is repeated, semantics become confusing;

Assign first, use later: this prevents obvious errors such as “x not defined” during interpretation or “x not allocated” during constraining.

Structs

LEM

A LEM has the name for the inputs and its characteristic control node

Var

Named references to be bound to Ptrs.

Enums

LEMCTL

The basic control nodes for LEM logical paths.

LEMOP

The atomic operations of LEMs.

Type Definitions

AString

AVec