Module lurk::lem::circuit

Constraint system for LEM

This module implements the generation of bellperson constraints for LEM, such that it can be used with Nova folding to prove evaluations of Lurk expressions.

“Concrete” and “virtual” paths

Control statements such as matches introduce branching, meaning execution can follow different paths depending on a value. The real path of execution for a particular instance we call the concrete path. The other paths which are not followed we call virtual paths. A mechanism to “relax” the constraints for the virtual paths while also properly enforcing the correctness for what happens on the concrete path is thus needed.

We do that by using implication gadgets. An implication of the form A → B is trivially true if A is false. But if A is true, then the implication is true iff B is true. In other words, B is irrelevant if A is false, which is the behavior we want for the virtual paths.

With that in mind, we can keep track of booleans that tell us whether we’re on a concrete or a virtual path and use such booleans as the premises to build the constraints we care about with implication gadgets.

Structs

GlobalAllocator

Manages global allocations for constants in a constraint system

Enums

AllocatedVal