Type Definition lurk::lem::pointers::ZPtr

pub type ZPtr<F> = ZPtr<Tag, F>;

A ZPtr is the result of “hydrating” a Ptr, which is a process of replacing indices by hashes. That is, a ZPtr is a content-addressed, tagged, pointer. By analogy, we can view ordinary field elements as hydrated raw pointers.

With ZPtrs we are able to content-address arbitrary DAGs, and thus be able to represent these data structures as field elements. This is how we can prove facts about data structures only using field elements. ZPtrs are also useful when we want to content-address the store.

In principle, ZPtrs could be used in place of Ptr, but it is important to note that content-addressing can be expensive, especially in the context of interpretation, because of the Poseidon hashes. That’s why we operate on Ptrs when interpreting LEMs and delay the need for ZPtrs as much as possible.

Implementations

impl<F: LurkField> ZPtr<F>

pub fn dummy() -> Self