Struct lurk::lem::LEM

pub struct LEM { /* private fields */ }

Lurk Evaluation Model (LEM)

A LEM is a description of Lurk’s evaluation algorithm, encoded as data. In other words, it’s a meta-representation of Lurk’s step function.

The motivation behind LEM is the fact that hand-writing the circuit is a fragile process that hinders experimentation and safety. Thus we would like to bootstrap the circuit automatically, given a higher level description of the step function.

LEM also allows the Store API to be completely abstracted away from the responsibilities of LEM authors. Indeed, we want the implementation details of the Store not to be important at LEM definition time.

Data semantics

A LEM describes how to handle pointers with “meta pointers”, which are basically named references. Instead of saying let foo ... in Rust, we use a MetaPtr("foo") in LEM.

The actual algorithm is encoded with a LEM operation (LEMOP). It’s worth noting that one of the LEM operators is in fact a vector of operators, which allows imperative expressiveness.

Interpretation

Running a LEM is done via interpretation, which might be a bit slower than calling Rust functions directly. But it also has its advantages:

The logic to collect data during execution can be factored out from the definition of the step function. This process is needed in order to evidence the inputs for the circuit at proving time;

Actually, such logic to collect data is a natural consequence of the fact that we’re on a higher level of abstraction. Relevant data is not simply stored on rust variables that die after the function ends. On the contrary, all relevant data lives on HashMaps that are also a product of the interpreted LEM.

Constraining

This is the process of creating the circuit, which we want to be done automatically for whoever creates a LEM. Each LEMOP has to be precisely constrained in such a way that the resulting circuits accepts a witness iff it was generated by a valid interpretation of the LEM at play.

Static checks of correctness

Since a LEM is an algorithm encoded as data, we can perform static checks of correctness as some form of (automated) formal verification. Here are some (WIP) properties we want a LEM to have before we can adopt it as a proper Lurk step function:

Static single assignments: overwriting meta pointers would erase relevant data needed to feed the circuit at proving time. We don’t want to lose any piece of information that the prover might know;

Non-duplicated input labels: right at the start of interpretation, the input labels are bound to the actual pointers that represent the expression, environment and continuation. If some label is repeated, it will fatally break property 1;

One return per LEM path: a LEM must always specify an output regardless of the logical path it takes at interpretation time, otherwise there would be a chance of the next step starting with an unknown input. Also, a LEM should not specify more than an output per logical path because it would risk setting conflicting constraints for the output;

Assign first, use later: this prevents obvious errors such as “x not defined” during interpretation or “x not allocated” during constraining.

Implementations

impl LEM

pub fn constrain<F: LurkField, CS: ConstraintSystem<F>>( &self, cs: &mut CS, store: &mut Store<F>, witness: &Witness<F> ) -> Result<()>

Create R1CS constraints for LEM given an evaluation witness.

As we find recursive (non-leaf) LEM operations, we stack them to be constrained later, using hash maps to manage viariables and pointers in a way we can reference allocated variables that were previously created.

Notably, we use a variable branch_path_info: Option<Boolean> to encode a three-way information:

If it’s None, it means that no logical branches have happened in the LEM so far, meaning that the evaluation algorithm must have gone through that operation. In this case, we use regular equality enforcements

If it’s Some(concrete_path), it means that we’re on a logical LEM branch, which might be virtual or concrete depending on the witness. A virtual path is one that wasn’t taken during evaluation and thus its witness pointers and variables weren’t bound. A concrete path means that evaluation took that path and the witness data should be complete. For virtual paths we need to create dummy bindings and relax the enforcements with implications whose premises are false. So, in the end, we use implications on both virtual and concrete paths to make sure that the circuit structure is always the same, independently of the witness. The premise is precicely concrete_path.

impl LEM

pub fn run<F: LurkField>( &self, input: [Ptr<F>; 3], store: &mut Store<F> ) -> Result<Witness<F>>

Interprets a LEM using a stack of operations to be popped and executed. It modifies a Store and assigns Ptrs to MetaPtrs as it goes.

pub fn eval<F: LurkField>( &self, expr: Ptr<F>, store: &mut Store<F> ) -> Result<Vec<Witness<F>>>

Calls run until the stop contidion is satisfied, using the output of one iteration as the input of the next one.

impl LEM

pub fn new(input: [&str; 3], lem_op: LEMOP) -> Result<LEM>

Instantiates a LEM with the appropriate checks and transformations to make sure that interpretation and constraining will be smooth.