Module lurk::proof

This module offers a connection the the backend proving engine of Lurk. Abstracted behind the Prover and Verifier traits, the proving engine has two instantiations:

the Groth16/SnarkPack proving system, implemented in the groth16 module

the Nova proving system, implemented in the nova module.

Modules

groth16

An adapter to a Groth16 proving system implementation.

nova

An adapter to a Nova proving system implementation.

supernova

An adapter to a SuperNova proving system implementation.

Traits

NIVCProver

Supertrait for Prover that also supports NIVC.

Provable

A trait for provable structures over a field F.

Prover

A trait for a prover that works with a field F.

PublicParameters

A trait representing the public parameters for a proving system.

Functions

verify_sequential_css

Verifies a sequence of constraint systems (CSs) for sequentiality & validity.