Trait lurk::proof::nova::CurveCycleEquipped

pub trait CurveCycleEquipped: LurkField {
    type EE1: EvaluationEngineTrait<Self::E1>;
    type EE2: EvaluationEngineTrait<Self::E2>;
    type E1: Engine<Base = <Self::E2 as Engine>::Scalar, Scalar = Self>;
    type E2: Engine<Base = <Self::E1 as Engine>::Scalar>;
}

This trait defines most of the requirements for programming generically over the supported Nova curve cycles (currently Pallas/Vesta and BN254/Grumpkin). It being pegged on the LurkField trait encodes that we do not expect more than one such cycle to be supported at a time for a given field.

Required Associated Types

type EE1: EvaluationEngineTrait<Self::E1>

Why the next 2 types?

In theory it would be sufficient to abstract over the two group types of the curve cycle, but in practice Nova is a bit idiosyncratic in the [nova::traits::evaluation::EvaluationEngineTrait<G>], (PCS) it uses on these (its multilinear IPA : [nova::provider::ipa_pc::EvaluationEngine<G>]) and that implementation requires an additional trait bound CommitmentKeyExtTrait for this type.

The following abstracts over curve cycle groups for which there exists an implementation of [nova::traits::evaluation::EvaluationEngineTrait<G>], encapsulating these idiosyncracies within Nova. a concrete implementation of an [nova::traits::evaluation::EvaluationEngineTrait<G>] for G1,

type EE2: EvaluationEngineTrait<Self::E2>

a concrete implementation of an [nova::traits::evaluation::EvaluationEngineTrait<G>] for G2,

type E1: Engine<Base = <Self::E2 as Engine>::Scalar, Scalar = Self>

The group type for the first curve in the cycle.

type E2: Engine<Base = <Self::E1 as Engine>::Scalar>

The group type for the second curve in the cycle.

Object Safety

This trait is not object safe.