Trait lurk::proof::Provable

pub trait Provable<F: LurkField> {
    // Required methods
    fn public_inputs(&self) -> Vec<F>;
    fn public_input_size(&self) -> usize;
    fn num_frames(&self) -> usize;
}

A trait for provable structures over a field F.

Required Methods

fn public_inputs(&self) -> Vec<F>

Returns the public inputs of the provable structure.

fn public_input_size(&self) -> usize

Returns the size of the public inputs.

fn num_frames(&self) -> usize

Returns the number of reduction frames in the provable structure.