Trait lurk::proof::RecursiveSNARKTrait

pub trait RecursiveSNARKTrait<'a, F: CurveCycleEquipped, C: Coprocessor<F> + 'a, M: MultiFrameTrait<'a, F, C>>
where
    Self: Sized,
{
    type PublicParams;
    type ProveOutput;
    type ExtraVerifyInput;
    type ErrorType;

    // Required methods
    fn prove_recursively(
        pp: &Self::PublicParams,
        z0: &[F],
        steps: Vec<M>,
        store: &'a M::Store,
        reduction_count: usize,
        lang: Arc<Lang<F, C>>
    ) -> Result<Self::ProveOutput, ProofError>;
    fn compress(self, pp: &Self::PublicParams) -> Result<Self, ProofError>;
    fn verify(
        &self,
        pp: &Self::PublicParams,
        z0: &[F],
        zi: &[F],
        extra: Self::ExtraVerifyInput
    ) -> Result<bool, Self::ErrorType>;

    // Provided method
    fn z0_secondary() -> Vec<<F::E2 as Engine>::Scalar> { ... }
}

Trait to abstract Nova and SuperNova proofs

Required Associated Types

type PublicParams

Associated type for public parameters

type ProveOutput

Main output of prove_recursively, encoding the actual proof

type ExtraVerifyInput

Extra input for verify to be defined as needed

type ErrorType

Type for error potentially thrown during verification

Required Methods

fn prove_recursively( pp: &Self::PublicParams, z0: &[F], steps: Vec<M>, store: &'a M::Store, reduction_count: usize, lang: Arc<Lang<F, C>> ) -> Result<Self::ProveOutput, ProofError>

Generate the recursive SNARK, encoded in ProveOutput

fn compress(self, pp: &Self::PublicParams) -> Result<Self, ProofError>

Compress a proof

fn verify( &self, pp: &Self::PublicParams, z0: &[F], zi: &[F], extra: Self::ExtraVerifyInput ) -> Result<bool, Self::ErrorType>

Verify the proof given the public parameters, the input and output values and the extra custom argument defined by who implements this trait.

Provided Methods

fn z0_secondary() -> Vec<<F::E2 as Engine>::Scalar>

Return the z0_secondary

Object Safety

This trait is not object safe.