Type Alias lurk::ptr::ContPtr

pub type ContPtr<F> = GPtr<F, ContTag>;

Aliased Type

struct ContPtr<F> {
    pub tag: ContTag,
    pub raw: RawPtr,
    pub _f: PhantomData<F>,
}

Fields
tag: ContTag

An expression tag

raw: RawPtr

The underlying pointer, which can be null, opaque, or an index

_f: PhantomData<F>

PhantomData is needed to consume the F: LurkField parameter, since we want to pin our Ptr to a specific field (even though we don’t actually use it)

Implementations

impl<F: LurkField> ContPtr<F>

pub fn new(tag: ContTag, raw: RawPtr) -> Self

pub const fn is_error(&self) -> bool