Type Alias lurk::ptr::Ptr

pub type Ptr<F> = GPtr<F, ExprTag>;

Aliased Type

struct Ptr<F> {
    pub tag: ExprTag,
    pub raw: RawPtr,
    pub _f: PhantomData<F>,
}

Fields
tag: ExprTag

An expression tag

raw: RawPtr

The underlying pointer, which can be null, opaque, or an index

_f: PhantomData<F>

PhantomData is needed to consume the F: LurkField parameter, since we want to pin our Ptr to a specific field (even though we don’t actually use it)

Implementations

impl<F: LurkField> Ptr<F>

pub fn allocate_maybe_fun_unconstrained<CS: ConstraintSystem<F>>( cs: CS, store: &Store<F>, maybe_fun: Option<&Ptr<F>> ) -> Result<(AllocatedNum<F>, AllocatedPtr<F>, AllocatedPtr<F>, AllocatedPtr<F>), SynthesisError>