Module lurk::public_parameters::instance

Instances

We manage the various kinds of public parameters of Nova and SuperNova with Instances. Each instance is a unique key representing a unique parameter object cached in the file system. The module provides the API to smoothly fetch/write/update params and sets of params.

Currently, [::nova] defines three “kinds” of parameter objects that need to be cached. The first kind corresponds to [::nova::PublicParams], which contain a full instance of public parameters required for a Nova proof. The second and third kinds correspond to [::nova::supernova::AuxParams] and [::nova::supernova::CircuitParams], which contain partial instances of public parameters required for a SuperNova proof.

Hashing

When inserting an (instance, parameter) pair into the cache, we compute an unique cache key for the instance to name the cached files. In this computation, we hash the unique identifying information of the instance.

For [::nova::PublicParams], this is the hash of the shape primary circuit

For [::nova::supernova::AuxParams], this is the combine hash of all of the primary circuit hashes of the SuperNova instance.

For [::nova::supernova::CircuitParams], this is the hash of the shape of its particular SuperNova circuit.

Because the first primary circuit of Lurk will always be the universal Lurk circuit, which may get very large at high rc, we instead hash an “ad-hoc” circuit that still uniquely represents the universal Lurk circuit, but is much smaller and faster to hash.

The idea is this: we have the large rc universal Lurk circuit as:

lurk-circuit(rc) = start bits -- frame -- glue -- frame -- glue ... -- end bits (repeat rc times)

Instead we hash an ad-hoc circuit (rc=2, which allows capturing our “glue”):

lurk-circuit(2) = start bits -- frame -- glue -- frame -- end bits

and hash (lurk-circuit(2), rc). This ensures that we include all the information about circuit plumbing done by start-bits, end-bits, and glue in the hash.

Structs

Instance

Instances compute a cache-key for the objects they represent and are themselves hashed to be put onto the file system. For each instance, we also have a Metadata json that contains useful information about it.

Metadata

What we put into the cache

Enums

Kind

From [::nova], there are 3 “kinds” of public param objects that need to be cached. The first kind corresponds to nova::NovaPublicParams, which contain a full instance of public parameters required for a Nova proof. The second and third kinds correspond to supernova::SuperNovaAuxParams and [supernova::SuperNovaCircuitParams], which contain partial instances of public parameters required for a SuperNova proof.