Struct lurk::store::Store

pub struct Store<F: LurkField> {

    pub cons_store: IndexSet<(Ptr<F>, Ptr<F>), RandomState>,
    pub comm_store: IndexSet<(FWrap<F>, Ptr<F>), RandomState>,
    pub fun_store: IndexSet<(Ptr<F>, Ptr<F>, Ptr<F>), RandomState>,
    pub sym_store: StringSet,
    pub num_store: IndexSet<Num<F>, RandomState>,
    pub str_store: StringSet,
    pub thunk_store: IndexSet<Thunk<F>, RandomState>,
    pub call0_store: IndexSet<(Ptr<F>, ContPtr<F>), RandomState>,
    pub call_store: IndexSet<(Ptr<F>, Ptr<F>, ContPtr<F>), RandomState>,
    pub call2_store: IndexSet<(Ptr<F>, Ptr<F>, ContPtr<F>), RandomState>,
    pub tail_store: IndexSet<(Ptr<F>, ContPtr<F>), RandomState>,
    pub lookup_store: IndexSet<(Ptr<F>, ContPtr<F>), RandomState>,
    pub unop_store: IndexSet<(Op1, ContPtr<F>), RandomState>,
    pub binop_store: IndexSet<(Op2, Ptr<F>, Ptr<F>, ContPtr<F>), RandomState>,
    pub binop2_store: IndexSet<(Op2, Ptr<F>, ContPtr<F>), RandomState>,
    pub if_store: IndexSet<(Ptr<F>, ContPtr<F>), RandomState>,
    pub let_store: IndexSet<(Ptr<F>, Ptr<F>, Ptr<F>, ContPtr<F>), RandomState>,
    pub letrec_store: IndexSet<(Ptr<F>, Ptr<F>, Ptr<F>, ContPtr<F>), RandomState>,
    pub emit_store: IndexSet<ContPtr<F>, RandomState>,
    pub opaque_ptrs: IndexSet<ScalarPtr<F>, RandomState>,
    pub opaque_cont_ptrs: IndexSet<ScalarContPtr<F>, RandomState>,
    pub scalar_ptr_map: DashMap<ScalarPtr<F>, Ptr<F>, RandomState>,
    pub scalar_ptr_cont_map: DashMap<ScalarContPtr<F>, ContPtr<F>, RandomState>,
    pub poseidon_cache: PoseidonCache<F>,
    pub dehydrated: Vec<Ptr<F>>,
    pub dehydrated_cont: Vec<ContPtr<F>>,
    pub pointer_scalar_ptr_cache: DashMap<Ptr<F>, ScalarPtr<F>>,
    pub lurk_package: Arc<Package>,
    pub constants: OnceCell<NamedConstants<F>>,

}

Fields
cons_store: IndexSet<(Ptr<F>, Ptr<F>), RandomState>comm_store: IndexSet<(FWrap<F>, Ptr<F>), RandomState>fun_store: IndexSet<(Ptr<F>, Ptr<F>, Ptr<F>), RandomState>sym_store: StringSetnum_store: IndexSet<Num<F>, RandomState>str_store: StringSetthunk_store: IndexSet<Thunk<F>, RandomState>call0_store: IndexSet<(Ptr<F>, ContPtr<F>), RandomState>call_store: IndexSet<(Ptr<F>, Ptr<F>, ContPtr<F>), RandomState>call2_store: IndexSet<(Ptr<F>, Ptr<F>, ContPtr<F>), RandomState>tail_store: IndexSet<(Ptr<F>, ContPtr<F>), RandomState>lookup_store: IndexSet<(Ptr<F>, ContPtr<F>), RandomState>unop_store: IndexSet<(Op1, ContPtr<F>), RandomState>binop_store: IndexSet<(Op2, Ptr<F>, Ptr<F>, ContPtr<F>), RandomState>binop2_store: IndexSet<(Op2, Ptr<F>, ContPtr<F>), RandomState>if_store: IndexSet<(Ptr<F>, ContPtr<F>), RandomState>let_store: IndexSet<(Ptr<F>, Ptr<F>, Ptr<F>, ContPtr<F>), RandomState>letrec_store: IndexSet<(Ptr<F>, Ptr<F>, Ptr<F>, ContPtr<F>), RandomState>emit_store: IndexSet<ContPtr<F>, RandomState>opaque_ptrs: IndexSet<ScalarPtr<F>, RandomState>opaque_cont_ptrs: IndexSet<ScalarContPtr<F>, RandomState>scalar_ptr_map: DashMap<ScalarPtr<F>, Ptr<F>, RandomState>

Holds a mapping of ScalarPtr -> Ptr for reverse lookups

scalar_ptr_cont_map: DashMap<ScalarContPtr<F>, ContPtr<F>, RandomState>

Holds a mapping of ScalarPtr -> ContPtr for reverse lookups

poseidon_cache: PoseidonCache<F>

Caches poseidon hashes

dehydrated: Vec<Ptr<F>>

Contains Ptrs which have not yet been hydrated.

dehydrated_cont: Vec<ContPtr<F>>pointer_scalar_ptr_cache: DashMap<Ptr<F>, ScalarPtr<F>>lurk_package: Arc<Package>constants: OnceCell<NamedConstants<F>>
Implementations

impl<F: LurkField> Store<F>

pub fn read(&mut self, input: &str) -> Result<Ptr<F>, Error>

pub fn read_in_package( &mut self, input: &str, package: &Package ) -> Result<Ptr<F>, Error>

pub fn read_string<T: Iterator<Item = char>>( &mut self, chars: &mut PeekMoreIterator<T> ) -> Result<Ptr<F>, Error>

pub fn read_maybe_meta<T: Iterator<Item = char>>( &mut self, chars: &mut PeekMoreIterator<T>, package: &Package ) -> Result<(Ptr<F>, bool), Error>

pub fn read_next<T: Iterator<Item = char>>( &mut self, chars: &mut PeekMoreIterator<T>, package: &Package ) -> Result<Ptr<F>, Error>

impl<F: LurkField> Store<F>

These methods provide a more ergonomic means of constructing and manipulating Lurk data. They can be thought of as a minimal DSL for working with Lurk data in Rust code. Prefer these methods when constructing literal data or assembling program fragments in tests or during evaluation, etc.

pub fn nil(&mut self) -> Ptr<F>

pub fn t(&mut self) -> Ptr<F>

pub fn cons(&mut self, car: Ptr<F>, cdr: Ptr<F>) -> Ptr<F>

pub fn strcons(&mut self, car: Ptr<F>, cdr: Ptr<F>) -> Ptr<F>

pub fn hidden(&self, secret: F, payload: Ptr<F>) -> Option<Ptr<F>>

pub fn hide(&mut self, secret: F, payload: Ptr<F>) -> Ptr<F>

pub fn open(&self, ptr: Ptr<F>) -> Option<(F, Ptr<F>)>

pub fn open_mut(&mut self, ptr: Ptr<F>) -> Result<(F, Ptr<F>), Error>

pub fn secret(&self, ptr: Ptr<F>) -> Option<Ptr<F>>

pub fn secret_mut(&mut self, ptr: Ptr<F>) -> Result<Ptr<F>, Error>

pub fn list(&mut self, elts: &[Ptr<F>]) -> Ptr<F>

pub fn num<T: Into<Num<F>>>(&mut self, num: T) -> Ptr<F>

pub fn uint64(&mut self, n: u64) -> Ptr<F>

pub fn str<T: AsRef<str>>(&mut self, name: T) -> Ptr<F>

pub fn lurk_sym<T: AsRef<str>>(&mut self, name: T) -> Ptr<F>

pub fn sym<T: AsRef<str>>(&mut self, name: T) -> Ptr<F>

pub fn key<T: AsRef<str>>(&mut self, name: T) -> Ptr<F>

pub fn root_sym<T: AsRef<str>>(&mut self, name: T, is_keyword: bool) -> Ptr<F>

pub fn car(&self, expr: &Ptr<F>) -> Result<Ptr<F>, Error>

pub fn cdr(&self, expr: &Ptr<F>) -> Result<Ptr<F>, Error>

pub fn new() -> Self

pub fn intern_nil(&mut self) -> Ptr<F>

pub fn get_nil(&self) -> Ptr<F>

pub fn get_begin(&self) -> Ptr<F>

pub fn get_quote(&self) -> Ptr<F>

pub fn get_t(&self) -> Ptr<F>

pub fn intern_cons(&mut self, car: Ptr<F>, cdr: Ptr<F>) -> Ptr<F>

pub fn intern_strcons(&mut self, car: Ptr<F>, cdr: Ptr<F>) -> Ptr<F>

pub fn intern_comm(&mut self, secret: F, payload: Ptr<F>) -> Ptr<F>

pub fn get_maybe_opaque(&self, tag: ExprTag, hash: F) -> Option<Ptr<F>>

pub fn intern_scalar_cont_ptr( &mut self, ptr: ScalarContPtr<F>, scalar_store: &ScalarStore<F> ) -> Option<ContPtr<F>>

pub fn intern_scalar_ptr( &mut self, scalar_ptr: ScalarPtr<F>, scalar_store: &ScalarStore<F> ) -> Option<Ptr<F>>

pub fn intern_maybe_opaque_fun(&mut self, hash: F) -> Ptr<F>

pub fn intern_maybe_opaque_sym(&mut self, hash: F) -> Ptr<F>

pub fn intern_maybe_opaque_cons(&mut self, hash: F) -> Ptr<F>

pub fn intern_maybe_opaque_comm(&mut self, hash: F) -> Ptr<F>

pub fn intern_opaque_fun(&mut self, hash: F) -> Ptr<F>

pub fn intern_opaque_sym(&mut self, hash: F) -> Ptr<F>

pub fn intern_opaque_cons(&mut self, hash: F) -> Ptr<F>

pub fn intern_opaque_comm(&mut self, hash: F) -> Ptr<F>

pub fn intern_list(&mut self, elts: &[Ptr<F>]) -> Ptr<F>

Helper to allocate a list, instead of manually using cons.

pub fn intern_sym_with_case_conversion<T: AsRef<str>>( &mut self, name: T, package: &Package ) -> Ptr<F>

pub fn intern_sym(&mut self, sym: &Sym) -> Ptr<F>

pub fn intern_key(&mut self, sym: &Sym) -> Ptr<F>

pub fn get_lurk_sym<T: AsRef<str>>( &self, name: T, convert_case: bool ) -> Option<Ptr<F>>

pub fn intern_num<T: Into<Num<F>>>(&mut self, num: T) -> Ptr<F>

pub fn get_num<T: Into<Num<F>>>(&self, num: T) -> Option<Ptr<F>>

pub fn get_char(&self, c: char) -> Ptr<F>

pub fn get_char_from_u32(&self, code: u32) -> Ptr<F>

pub fn get_u64(&self, n: u64) -> Ptr<F>

pub fn intern_str<T: AsRef<str>>(&mut self, str: T) -> Ptr<F>

pub fn get_str<T: AsRef<str>>(&self, name: T) -> Option<Ptr<F>>

pub fn get_sym(&self, sym: &Sym) -> Option<Ptr<F>>

pub fn intern_fun( &mut self, arg: Ptr<F>, body: Ptr<F>, closed_env: Ptr<F> ) -> Ptr<F>

pub fn intern_thunk(&mut self, thunk: Thunk<F>) -> Ptr<F>

pub fn mark_dehydrated_cont(&mut self, p: ContPtr<F>) -> ContPtr<F>

pub fn get_cont_outermost(&self) -> ContPtr<F>

pub fn get_cont_error(&self) -> ContPtr<F>

pub fn get_cont_terminal(&self) -> ContPtr<F>

pub fn get_cont_dummy(&self) -> ContPtr<F>

pub fn intern_cont_error(&mut self) -> ContPtr<F>

pub fn intern_cont_outermost(&mut self) -> ContPtr<F>

pub fn intern_cont_terminal(&mut self) -> ContPtr<F>

pub fn intern_cont_dummy(&mut self) -> ContPtr<F>

pub fn scalar_from_parts(&self, tag: F, value: F) -> Option<ScalarPtr<F>>

pub fn scalar_from_parts_cont( &self, tag: F, value: F ) -> Option<ScalarContPtr<F>>

pub fn fetch_scalar(&self, scalar_ptr: &ScalarPtr<F>) -> Option<Ptr<F>>

pub fn fetch_scalar_cont( &self, scalar_ptr: &ScalarContPtr<F> ) -> Option<ContPtr<F>>

pub fn fetch_maybe_sym(&self, ptr: &Ptr<F>) -> Option<Sym>

pub fn fetch_sym(&self, ptr: &Ptr<F>) -> Option<Sym>

pub fn fetch_str(&self, ptr: &Ptr<F>) -> Option<&str>

pub fn fetch_char(&self, ptr: &Ptr<F>) -> Option<char>

pub fn fetch_fun(&self, ptr: &Ptr<F>) -> Option<&(Ptr<F>, Ptr<F>, Ptr<F>)>

pub fn fetch_cons(&self, ptr: &Ptr<F>) -> Option<&(Ptr<F>, Ptr<F>)>

pub fn fetch_comm(&self, ptr: &Ptr<F>) -> Option<&(FWrap<F>, Ptr<F>)>

pub fn fetch_num(&self, ptr: &Ptr<F>) -> Option<&Num<F>>

pub fn fetch_uint(&self, ptr: &Ptr<F>) -> Option<UInt>

pub fn fetch(&self, ptr: &Ptr<F>) -> Option<Expression<'_, F>>

pub fn fetch_list(&self, ptr: &Ptr<F>) -> Option<Vec<Ptr<F>>>

Returns a Vec of Ptrs representing the elements of a proper list, ptr. This is intended to be the inverse of Store::list(). IF ptr isn’t a proper list, return None.

pub fn fetch_cont(&self, ptr: &ContPtr<F>) -> Option<Continuation<F>>

pub fn car_cdr_mut(&mut self, ptr: &Ptr<F>) -> Result<(Ptr<F>, Ptr<F>), Error>

Mutable version of car_cdr to handle Str. (cdr str) may return a new str (the tail), which must be allocated.

pub fn car_cdr(&self, ptr: &Ptr<F>) -> Result<(Ptr<F>, Ptr<F>), Error>

pub fn hash_expr(&self, ptr: &Ptr<F>) -> Option<ScalarPtr<F>>

pub fn get_expr_hash(&self, ptr: &Ptr<F>) -> Option<ScalarPtr<F>>

pub fn hash_expr_aux( &self, ptr: &Ptr<F>, mode: HashScalar ) -> Option<ScalarPtr<F>>

pub fn hash_cont(&self, ptr: &ContPtr<F>) -> Option<ScalarContPtr<F>>

pub fn get_hash_components_cont(&self, ptr: &ContPtr<F>) -> Option<[F; 8]>

pub fn get_hash_components_thunk(&self, thunk: &Thunk<F>) -> Option<[F; 4]>

pub fn get_opaque_ptr(&self, ptr: Ptr<F>) -> Option<ScalarPtr<F>>

pub fn hash_sym(&self, sym: Ptr<F>, mode: HashScalar) -> Option<ScalarPtr<F>>

pub fn hash_string_mut<T: AsRef<str>>(&mut self, s: T) -> F

pub fn hash_nil(&self, mode: HashScalar) -> Option<ScalarPtr<F>>

pub fn new_opaque_ptr(&mut self) -> Ptr<F>

pub fn ptr_eq(&self, a: &Ptr<F>, b: &Ptr<F>) -> Result<bool, Error>

pub fn cons_eq(&self, a: &Ptr<F>, b: &Ptr<F>) -> bool

pub fn hydrate_scalar_cache(&mut self)

Fill the cache for Scalars. Only Ptrs which have been interned since last hydration will be hashed, so it is safe to call this incrementally. However, for best proving performance, we should call exactly once so all hashing can be batched, e.g. on the GPU.

pub fn get_constants(&self) -> &NamedConstants<F>

pub fn intern_sym_and_ancestors(&mut self, sym: &Sym) -> Option<Ptr<F>>