Module lurk::z_data

ZData is a lightweight binary data serialization format.

A ZData value is a n-ary tree with two types of nodes:

Atom: a leaf node containing a byte sequence

Cell: an interior node containing an arbitrary number of child nodes

ZData values are encoded as a sequence of bytes in a compact binary format.

This module also provides several traits that can be used to encode and decode Rust types into ZData values.

Modules

z_cont

z_expr

z_ptr

z_store

Enums

ZData

ZData is a binary tree with two types of nodes: Atom and Cell.

Functions

from_z_data

to_z_data