Enum lurk::z_data::z_expr::ZExpr

pub enum ZExpr<F: LurkField> {

    Nil,
    Cons(ZExprPtr<F>, ZExprPtr<F>),
    Comm(F, ZExprPtr<F>),
    RootSym,
    Sym(ZExprPtr<F>, ZExprPtr<F>),
    Key(ZExprPtr<F>, ZExprPtr<F>),
    Fun {
        arg: ZExprPtr<F>,
        body: ZExprPtr<F>,
        closed_env: ZExprPtr<F>,
    },
    Num(F),
    EmptyStr,
    Str(ZExprPtr<F>, ZExprPtr<F>),
    Thunk(ZExprPtr<F>, ZContPtr<F>),
    Char(char),
    UInt(UInt),

}

A ZExpr is the content-addressed representation of a Lurk expression, which enables efficient serialization and sharing of hashed Lurk data via associated ZExprPtrs.

Variants

Nil

A null expression

Cons(ZExprPtr<F>, ZExprPtr<F>)

A cons list of ZExprPtrs

Comm(F, ZExprPtr<F>)

A commitment, which contains an opaque value and a pointer to the hidden data in the ZStore

RootSym

Sym(ZExprPtr<F>, ZExprPtr<F>)

Contains a symbol (a list of strings) and a pointer to the tail.

Key(ZExprPtr<F>, ZExprPtr<F>)

Fun

Fields

arg: ZExprPtr<F>

body: ZExprPtr<F>

closed_env: ZExprPtr<F>

Num(F)

A field element representing a number

EmptyStr

Str(ZExprPtr<F>, ZExprPtr<F>)

Contains a string and a pointer to the tail.

Thunk(ZExprPtr<F>, ZContPtr<F>)

An unevaluated expression and continuation

Char(char)

UInt(UInt)

Implementations

impl<F: LurkField> ZExpr<F>

pub fn z_ptr(&self, cache: &PoseidonCache<F>) -> ZExprPtr<F>

Constructs a ZExprPtr from a ZExpr, creating a Poseidon hash from the consituent elements if needed

pub fn from_ptr(store: &Store<F>, ptr: &Ptr<F>) -> Option<Self>

Constructs a ZExpr by fetching ptr’s expression from the store and hashing it