Crate lurk

Re-exports

pub use symbol::Symbol;

pub use uint::UInt;

pub use z_data::z_cont;

pub use z_data::z_expr;

pub use z_data::z_ptr;

pub use z_data::z_store;

Modules

cache_map

circuit

cont

coprocessor

error

eval

expr

field

The finite field used in the language.

hash

hash_witness

lem

Lurk Evaluation Model (LEM)

parser

proof

This module offers a connection the the backend proving engine of Lurk. Abstracted behind the Prover and Verifier traits, the proving engine has two instantiations:

ptr

public_parameters

repl

store

symbol

syntax

tag

uint

writer

z_data

ZData is a lightweight binary data serialization format.

Macros

char

keyword

lem

lem_code

lemop

list

lurksym

num

str

sym

symbol

uint

var

vars

Enums

Num

Finite field element type for Lurk. Has different internal representations to optimize evaluation.

Constants

TEST_SEED