Trait lurk::coprocessor::Coprocessor

pub trait Coprocessor<F: LurkField>: Clone + Debug + Sync + Send + CoCircuit<F> {
    // Required methods
    fn eval_arity(&self) -> usize;
    fn evaluate_simple(&self, _s: &Store<F>, _args: &[Ptr]) -> Ptr;

    // Provided methods
    fn has_circuit(&self) -> bool { ... }
    fn evaluate_internal(&self, s: &Store<F>, ptrs: &[Ptr]) -> Vec<Ptr> { ... }
    fn evaluate(
        &self,
        s: &Store<F>,
        args: &[Ptr],
        env: &Ptr,
        cont: &Ptr
    ) -> Vec<Ptr> { ... }
}

Coprocessor is a trait that represents a generalized interface for coprocessors. Coprocessors augment the Lurk circuit and evaluation with additional built-in functionality. This trait generalizes over functionality needed in the evaluator, the sibling CoCircuit trait, generalizes over functionality needed in the circuit.

The trait is implemented by concrete coprocessor types, such as DummyCoprocessor.

Pattern

We use a type class (trait) pattern to provide extensibility for the coprocessor implementations. The pattern involves:

A trait crate::coprocessor::Coprocessor, which defines the methods and behavior for all coprocessors.

An enum such as crate::eval::lang::Coproc, which “closes” the hierarchy of possible coprocessor implementations we want to instantiate at a particular point in the code.

Required Methods

fn eval_arity(&self) -> usize

fn evaluate_simple(&self, _s: &Store<F>, _args: &[Ptr]) -> Ptr

Provided Methods

fn has_circuit(&self) -> bool

Returns true if this Coprocessor actually implements a circuit.

fn evaluate_internal(&self, s: &Store<F>, ptrs: &[Ptr]) -> Vec<Ptr>

Function for internal plumbing. Reimplementing is not recommended

fn evaluate( &self, s: &Store<F>, args: &[Ptr], env: &Ptr, cont: &Ptr ) -> Vec<Ptr>

Object Safety

This trait is not object safe.