Trait lurk::field::LurkField

pub trait LurkField: PrimeField + PrimeFieldBits {
    const FIELD: LanguageField;

    // Provided methods
    fn to_bytes(self) -> Vec<u8>  { ... }
    fn from_bytes(bs: &[u8]) -> Option<Self> { ... }
    fn hex_digits(self) -> String { ... }
    fn trimmed_hex_digits(self) -> String { ... }
    fn to_u16(&self) -> Option<u16> { ... }
    fn to_u32(&self) -> Option<u32> { ... }
    fn to_char(&self) -> Option<char> { ... }
    fn to_u64(&self) -> Option<u64> { ... }
    fn to_u32_unchecked(&self) -> u32 { ... }
    fn to_u64_unchecked(&self) -> u64 { ... }
    fn from_u64(x: u64) -> Self { ... }
    fn from_u32(x: u32) -> Self { ... }
    fn from_u16(x: u16) -> Self { ... }
    fn from_char(x: char) -> Self { ... }
    fn most_negative() -> Self { ... }
    fn most_positive() -> Self { ... }
    fn is_negative(&self) -> bool { ... }
    fn from_expr_tag(tag: ExprTag) -> Self { ... }
    fn to_expr_tag(&self) -> Option<ExprTag> { ... }
    fn from_cont_tag(tag: ContTag) -> Self { ... }
    fn to_cont_tag(&self) -> Option<ContTag> { ... }
    fn from_op1(tag: Op1) -> Self { ... }
    fn to_op1(&self) -> Option<Op1> { ... }
    fn from_op2(tag: Op2) -> Self { ... }
    fn to_op2(&self) -> Option<Op2> { ... }
    fn get_field(&self) -> LanguageField { ... }

}

Trait implemented by finite fields used in the language

Required Associated Constants

const FIELD: LanguageField

The type of the field element’s representation

Provided Methods

fn to_bytes(self) -> Vec<u8>

Converts the field element to a byte vector

fn from_bytes(bs: &[u8]) -> Option<Self>

Attempts to construct a field element from a byte slice

fn hex_digits(self) -> String

Converts the field element to a hexadecimal string

fn trimmed_hex_digits(self) -> String

Converts the field to a variable-length hex string

fn to_u16(&self) -> Option<u16>

Attempts to convert the field element to a u16

fn to_u32(&self) -> Option<u32>

Attempts to convert the field element to a u32

fn to_char(&self) -> Option<char>

Attempts to convert the field element to a char

fn to_u64(&self) -> Option<u64>

Attempts to convert the field element to a u64

fn to_u32_unchecked(&self) -> u32

Converts the first 4 bytes of the field element to a u32

fn to_u64_unchecked(&self) -> u64

Converts the first 8 bytes of the field element to a u64

fn from_u64(x: u64) -> Self

Constructs a field element from a u64

fn from_u32(x: u32) -> Self

Constructs a field element from a u32

fn from_u16(x: u16) -> Self

Constructs a field element from a u16

fn from_char(x: char) -> Self

Constructs a field element from a char

fn most_negative() -> Self

We define this to be the smallest negative field element

fn most_positive() -> Self

0 - 1 is one minus the modulus, which must be even in a prime field. The result is the largest field element which is even when doubled. We define this to be the most positive field element.

fn is_negative(&self) -> bool

A field element is defined to be negative if it is odd after doubling.

fn from_expr_tag(tag: ExprTag) -> Self

Constructs a field element from an ExprTag

fn to_expr_tag(&self) -> Option<ExprTag>

Attempts to convert the field element to an ExprTag

fn from_cont_tag(tag: ContTag) -> Self

Constructs a field element from a ContTag

fn to_cont_tag(&self) -> Option<ContTag>

Attempts to convert the field element to a ContTag

fn from_op1(tag: Op1) -> Self

Constructs a field element from an Op1

fn to_op1(&self) -> Option<Op1>

Attempts to convert the field element to an Op1

fn from_op2(tag: Op2) -> Self

Constructs a field element from an Op2

fn to_op2(&self) -> Option<Op2>

Attempts to convert the field element to an Op2

fn get_field(&self) -> LanguageField

Returns the LanguageField of the field