Module lurk::lem

Lurk Evaluation Model (LEM)

LEM is a simple, first order, referentially transparent language, designed to allow writing Lurk’s step function and Lurk coprocessors in a convenient way.

The motivation behind LEM is the fact that hand-writing the circuit is a fragile process that hinders experimentation and safety. Thus we would like to bootstrap the circuit automatically, as well as an interpretation algorithm that computes all non-deterministic advices for the circuit, given a higher level description of the step function.

LEM also allows the Store API to be completely abstracted away from the responsibilities of LEM authors. Indeed, we want the implementation details of the Store not to be important at LEM definition time.

Data semantics

The main data type that represents LEM code is the Func type. A Func is much like a function: it contains input parameters, output size, and a function body. The body of a function is a Block which is a sequence of operations Op followed by a control Ctrl statement.

Operations are much like let statements in functional languages. For example, a Op::Cons2(x, t, ys) is to be understood as let x = cons2(ys). If a second operation binds its result to the same variable as a previous operation, we shadow the previous value. There is no mutation, thus the language is referentially transparent.

A control statement is either a Return(xs), which exits a function and returns the values of the specified variables, or a match statement, which will do case analysis on the value of a variable and select the appropriate block to continue to.

Interpretation

The interpreter runs a LEM function given input values. Interpreting a LEM function will compute the values of each variable in the path of execution. In particular, it will compute all the non-deterministic advices that are needed to solve the circuit.

Synthesizing

Synthesizing is the process of building the circuit and solving it for a particular instance (i.e. finding a witness). All valid LEM functions can be synthesized if they were previously interpreted.

Code transformations and static checks of correctness

Here are some simple transformations and static checks of correctness we want to perform on a LEM function before interpreting and synthesizing it

All variables must be bound, no variable can be used before being bound

All returns within a block must be of the same size and equal to the function’s output size

Function calls must have the correct number of arguments and must bind the correct number of variables

No match statements should have conflicting cases

LEM should be transformed to SSA to make it simple to synthesize

We also check for variables that are not used. If intended they should be prefixed by “_”

Modules

circuit

Constraint system for LEM

eval

interpreter

multiframe

pointers

store

zstore

Structs

Block

A block is a sequence of operations followed by a control. Each block delimits their variables’ scope.

Func

A Func is a LEM function. It consist of input params, output size and a function body, which is a Block

Var

LEM variables

Enums

Ctrl

The basic control nodes for LEM logical paths.

Lit

LEM literals

Op

The atomic operations of LEMs.

Tag

LEM tags

Type Aliases

AString