Module lurk::parser::string

Enums

StringFragment

A string fragment contains a fragment of a string being parsed: either a non-empty Literal (a series of non-escaped characters), a single parsed escaped character, or a block of escaped whitespace.

Functions

parse_escaped_char

Parse an escaped character: \n, \t, \r, \u{00AC}, etc.

parse_escaped_whitespace

Parse a backslash, followed by any amount of whitespace. This is used later to discard any escaped whitespace.

parse_fragment

Combine parse_literal, parse_escaped_whitespace, and parse_escaped_char into a StringFragment.

parse_literal

parse_string

Parse a string with the outer delimiter characters

parse_string_inner

Parse a string. Use a loop of parse_fragment and push all of the fragments into an output string.

parse_string_inner1

Parse a non-empty string. Use a loop of parse_fragment and push all of the fragments into an output string.

parse_unicode

Parse a unicode sequence, of the form u{XXXX}, where XXXX is 1 to 6 hexadecimal numerals. We will combine this later with parse_escaped_char to parse sequences like \u{00AC}.