Module lurk::proof

This module offers a connection the the backend proving engine of Lurk. Abstracted behind the Prover and Verifier traits, the proving engine has two instantiations:

the Nova proving system, implemented in the nova module.

the SuperNova proving system, implemented in the supernova module.

Modules

nova

An adapter to a Nova proving system implementation.

supernova

An adapter to a SuperNova proving system implementation.

Traits

CEKState

The State of a CEK machine.

EvaluationStore

A trait for a store of expressions

FrameLike

A Frame of evaluation in a CEK machine.

MultiFrameTrait

Trait to support multiple MultiFrame implementations.

NIVCProver

Supertrait for Prover that also supports NIVC.

Provable

A trait for provable structures over a field F.

Prover

A trait for a prover that works with a field F.