Module lurk::proof::nova

An adapter to a Nova proving system implementation.

Structs

NovaProver

A struct for the Nova prover that operates on field elements of type F.

PublicParams

A struct that contains public parameters for the Nova proving system.

Enums

Proof

An enum representing the two types of proofs that can be generated and verified.

Traits

CurveCycleEquipped

This trait defines most of the requirements for programming generically over the supported Nova curve cycles (currently Pallas/Vesta and BN254/Grumpkin). It being pegged on the LurkField trait encodes that we do not expect more than one such cycle to be supported at a time for a given field.

Functions

circuit_cache_key

Computes a cache key of the primary circuit. The point is that if a circuit changes in any way but has the same rc/Lang, then we still want the public params to stay in sync with the changes.

circuits

Generates the circuits for the Nova proving system.

public_params

Generates the public parameters for the Nova proving system.

Type Aliases

C1LEM

LEM’s version of C1

C1Lurk

Type alias for a MultiFrame with S1 field elements. This uses the <::G1 as Group>::Scalar type for the G1 scalar field elements to reflect it this should not be used outside the Nova context

C2

Type alias for a Trivial Test Circuit with G2 scalar field elements.

EE1

Type alias for the Evaluation Engine using G1 group elements.

EE2

Type alias for the Evaluation Engine using G2 group elements.

G1

Convenience alias for the primary group type pegged to a LurkField through a CurveCycleEquipped type.

G2

Convenience alias for the secondary group type pegged to a LurkField through a CurveCycleEquipped type.

NovaCircuitShape

Type alias for Nova Circuit Parameters with the curve cycle types defined above.

NovaPublicParams

Type alias for Nova Public Parameters with the curve cycle types defined above.

SS1

Type alias for the Relaxed R1CS Spartan SNARK using G1 group elements, EE1.

SS2

Type alias for the Relaxed R1CS Spartan SNARK using G2 group elements, EE2.