Trait lurk::proof::nova::CurveCycleEquipped

pub trait CurveCycleEquipped: LurkField {
    type CK1: CommitmentKeyExtTrait<Self::G1>;
    type CK2: CommitmentKeyExtTrait<Self::G2>;
    type CE1: CommitmentEngineTrait<Self::G1, CommitmentKey = Self::CK1>;
    type CE2: CommitmentEngineTrait<Self::G2, CommitmentKey = Self::CK2>;
    type G1: Group<Base = <Self::G2 as Group>::Scalar, Scalar = Self, CE = Self::CE1>;
    type G2: Group<Base = <Self::G1 as Group>::Scalar, CE = Self::CE2>;
}

This trait defines most of the requirements for programming generically over the supported Nova curve cycles (currently Pallas/Vesta and BN254/Grumpkin). It being pegged on the LurkField trait encodes that we do not expect more than one such cycle to be supported at a time for a given field.

Required Associated Types

type CK1: CommitmentKeyExtTrait<Self::G1>

Why the next 4 types?

The next 4 types are purely technical, and aim at laying out type bounds in a way that rust can find them. They should eventually be replaceable by a bound on projections, once bounds on associated types progress. They are technically equivalent to bounds of <Self::G1::CE as CommitmentEngineTraitSelf::G1>::CommitmentKey: CommitmentKeyExtTrait<Self::G1, CE = <Self::G1 as Group>::CE>, <Self::G2::CE as CommitmentEngineTraitSelf::G2>::CommitmentKey: CommitmentKeyExtTrait<Self::G2, CE = ::CE>, but where clauses can’t be found by the compiler at the point where Self::G1, Self::G2 are used

OK, but why do we need bounds at all in the first place?

As to why those see https://github.com/microsoft/Nova/pull/200 and the bound CommitmentKey<G>: CommitmentKeyExtTrait<G, CE = G::CE> on [nova::provider::ipa_pc::EvaluationEngine<G>] Essentially, Nova relies on a commitment scheme that is additively homomorphic, but encodes the practicalities of this (properties are unwieldy to encode) in the form of this CommitmentKeyExtTrait. The type of the commitment key used for points of the first curve in the cycle.

type CK2: CommitmentKeyExtTrait<Self::G2>

The type of the commitment key used for points of the second curve in the cycle.

type CE1: CommitmentEngineTrait<Self::G1, CommitmentKey = Self::CK1>

The commitment engine type for the first curve in the cycle.

type CE2: CommitmentEngineTrait<Self::G2, CommitmentKey = Self::CK2>

The commitment engine type for the second curve in the cycle.

type G1: Group<Base = <Self::G2 as Group>::Scalar, Scalar = Self, CE = Self::CE1>

The group type for the first curve in the cycle.

type G2: Group<Base = <Self::G1 as Group>::Scalar, CE = Self::CE2>

The group type for the second curve in the cycle.