Type Alias lurk::proof::nova::C1

pub type C1<'a, F, C> = MultiFrame<'a, F, C>;

Type alias for a MultiFrame with S1 field elements. This uses the <::G1 as Group>::Scalar type for the G1 scalar field elements to reflect it this should not be used outside the Nova context

Aliased Type

struct C1<'a, F, C> {
    pub store: Option<&'a Store<F>>,
    pub input: Option<IO<F>>,
    pub output: Option<IO<F>>,
    pub frames: Option<Vec<CircuitFrame<'a, F, C>, Global>>,
    pub cached_witness: Option<WitnessCS<F>>,
    pub count: usize,
    pub folding_config: Arc<FoldingConfig<F, C>, Global>,
    pub meta: Meta<F>,
}

Fields
store: Option<&'a Store<F>>input: Option<IO<F>>output: Option<IO<F>>frames: Option<Vec<CircuitFrame<'a, F, C>, Global>>cached_witness: Option<WitnessCS<F>>count: usizefolding_config: Arc<FoldingConfig<F, C>, Global>meta: Meta<F>
Implementations

impl<'a, F: LurkField, C: Coprocessor<F>> MultiFrame<'a, F, C>

pub fn blank(folding_config: Arc<FoldingConfig<F, C>>, meta: Meta<F>) -> Self

pub fn get_store(&self) -> &Store<F>

pub fn from_frames( count: usize, frames: &[Frame<IO<F>, Witness<F>, F, C>], store: &'a Store<F>, folding_config: Arc<FoldingConfig<F, C>> ) -> Vec<Self>

pub fn synthesize_frames<CS: ConstraintSystem<F>>( &self, cs: &mut CS, store: &Store<F>, input_expr: AllocatedPtr<F>, input_env: AllocatedPtr<F>, input_cont: AllocatedContPtr<F>, frames: &[CircuitFrame<'_, F, C>], g: &GlobalAllocations<F> ) -> (AllocatedPtr<F>, AllocatedPtr<F>, AllocatedContPtr<F>)

pub fn synthesize_frames_sequential<CS: ConstraintSystem<F>>( &self, cs: &mut CS, store: &Store<F>, input_expr: AllocatedPtr<F>, input_env: AllocatedPtr<F>, input_cont: AllocatedContPtr<F>, frames: &[CircuitFrame<'_, F, C>], cons_and_cont_witnesses: Option<Vec<(ConsCircuitWitness<F>, ContCircuitWitness<F>)>>, g: &GlobalAllocations<F> ) -> (AllocatedPtr<F>, AllocatedPtr<F>, AllocatedContPtr<F>)

pub fn synthesize_frames_parallel<CS: ConstraintSystem<F>>( &self, cs: &mut CS, store: &Store<F>, input_expr: AllocatedPtr<F>, input_env: AllocatedPtr<F>, input_cont: AllocatedContPtr<F>, frames: &[CircuitFrame<'_, F, C>], g: &GlobalAllocations<F> ) -> (AllocatedPtr<F>, AllocatedPtr<F>, AllocatedContPtr<F>)

impl<F: LurkField, C: Coprocessor<F>> MultiFrame<'_, F, C>

pub fn precedes(&self, maybe_next: &Self) -> bool

impl<C: Coprocessor<Scalar>> MultiFrame<'_, <Bls12 as Engine>::Fr, C>

pub fn verify_groth16_proof( self, pvk: &PreparedVerifyingKey<Bls12>, proof: &Proof<Bls12> ) -> Result<bool, SynthesisError>

Verify a Groth16 Lurk proof.

impl<'a, F: LurkField, C: Coprocessor<F>> MultiFrame<'a, F, C>

pub fn circuit_index(&self) -> usize

Return the circuit index assigned to this MultiFrame’s inner computation, as labeled by its Meta, and determined by its FoldingConfig.