Trait lurk::proof::MultiFrameTrait

pub trait MultiFrameTrait<'a, F: LurkField, C: Coprocessor<F> + 'a>: Provable<F> + Circuit<F> + StepCircuit<F> + 'a {
    type Ptr: Debug + Eq + Copy;
    type ContPtr: Debug + Eq + Copy;
    type Store: Send + Sync + EvaluationStore<Ptr = Self::Ptr, ContPtr = Self::ContPtr>;
    type StoreError: Into<ProofError>;
    type EvalFrame: FrameLike<Self::Ptr, Self::ContPtr>;
    type CircuitFrame: FrameLike<Self::Ptr, Self::ContPtr, FrameIO = <Self::EvalFrame as FrameLike<Self::Ptr, Self::ContPtr>>::FrameIO>;
    type GlobalAllocation;
    type AllocatedIO;

    // Required methods
    fn emitted(
        store: &Self::Store,
        eval_frame: &Self::EvalFrame
    ) -> Vec<Self::Ptr>;
    fn significant_frame_count(frames: &[Self::EvalFrame]) -> usize;
    fn get_evaluation_frames(
        padding_predicate: impl Fn(usize) -> bool,
        expr: Self::Ptr,
        env: Self::Ptr,
        store: &Self::Store,
        limit: usize,
        land: &Lang<F, C>
    ) -> Result<Vec<Self::EvalFrame>, ProofError>;
    fn io_to_scalar_vector(
        store: &Self::Store,
        io: &<Self::EvalFrame as FrameLike<Self::Ptr, Self::ContPtr>>::FrameIO
    ) -> Result<Vec<F>, Self::StoreError>;
    fn precedes(&self, maybe_next: &Self) -> bool;
    fn compute_witness(&self, s: &Self::Store) -> WitnessCS<F>;
    fn cached_witness(&mut self) -> &mut Option<WitnessCS<F>>;
    fn output(
        &self
    ) -> &Option<<Self::EvalFrame as FrameLike<Self::Ptr, Self::ContPtr>>::FrameIO>;
    fn frames(&self) -> Option<&Vec<Self::CircuitFrame>>;
    fn synthesize_frames<CS: ConstraintSystem<F>>(
        &self,
        cs: &mut CS,
        store: &Self::Store,
        input: Self::AllocatedIO,
        frames: &[Self::CircuitFrame],
        g: &Self::GlobalAllocation
    ) -> Result<Self::AllocatedIO, SynthesisError>;
    fn blank(
        folding_config: Arc<FoldingConfig<F, C>>,
        meta: Meta<F>,
        pc: usize
    ) -> Self;
    fn from_frames(
        reduction_count: usize,
        frames: &[Self::EvalFrame],
        store: &'a Self::Store,
        folding_config: Arc<FoldingConfig<F, C>>
    ) -> Vec<Self>;
    fn make_dummy(
        reduction_count: usize,
        circuit_frame: Option<Self::CircuitFrame>,
        store: &'a Self::Store,
        folding_config: Arc<FoldingConfig<F, C>>,
        meta: Meta<F>
    ) -> Self;

}

Trait to support multiple MultiFrame implementations.

Required Associated Types

type Ptr: Debug + Eq + Copy

The associated Ptr type

type ContPtr: Debug + Eq + Copy

The associated ContPtr type

type Store: Send + Sync + EvaluationStore<Ptr = Self::Ptr, ContPtr = Self::ContPtr>

The associated Store type

type StoreError: Into<ProofError>

The error type for the Store type

type EvalFrame: FrameLike<Self::Ptr, Self::ContPtr>

The associated Frame type

type CircuitFrame: FrameLike<Self::Ptr, Self::ContPtr, FrameIO = <Self::EvalFrame as FrameLike<Self::Ptr, Self::ContPtr>>::FrameIO>

The associated CircuitFrame type

type GlobalAllocation

The associated type which manages global allocations

type AllocatedIO

The associated type of allocated input and output to the circuit

Required Methods

fn emitted(store: &Self::Store, eval_frame: &Self::EvalFrame) -> Vec<Self::Ptr>

the emitted frames

fn significant_frame_count(frames: &[Self::EvalFrame]) -> usize

Counting the number of non-trivial frames in the evaluation

fn get_evaluation_frames( padding_predicate: impl Fn(usize) -> bool, expr: Self::Ptr, env: Self::Ptr, store: &Self::Store, limit: usize, land: &Lang<F, C> ) -> Result<Vec<Self::EvalFrame>, ProofError>

Evaluates and generates the frames of the computation given the expression, environment, and store (IVC only, TODO NIVC)

fn io_to_scalar_vector( store: &Self::Store, io: &<Self::EvalFrame as FrameLike<Self::Ptr, Self::ContPtr>>::FrameIO ) -> Result<Vec<F>, Self::StoreError>

Returns a public IO vector when equipped with the local store, and the Self::Frame’s IO

fn precedes(&self, maybe_next: &Self) -> bool

Returns true if the supplied instance directly precedes this one in a sequential computation trace.

fn compute_witness(&self, s: &Self::Store) -> WitnessCS<F>

Populates a WitnessCS with the witness values for the given store.

fn cached_witness(&mut self) -> &mut Option<WitnessCS<F>>

Returns a reference to the cached witness values

fn output( &self ) -> &Option<<Self::EvalFrame as FrameLike<Self::Ptr, Self::ContPtr>>::FrameIO>

The output of the last frame

fn frames(&self) -> Option<&Vec<Self::CircuitFrame>>

Iterates through the Self::CircuitFrame instances

fn synthesize_frames<CS: ConstraintSystem<F>>( &self, cs: &mut CS, store: &Self::Store, input: Self::AllocatedIO, frames: &[Self::CircuitFrame], g: &Self::GlobalAllocation ) -> Result<Self::AllocatedIO, SynthesisError>

Synthesize some frames.

fn blank( folding_config: Arc<FoldingConfig<F, C>>, meta: Meta<F>, pc: usize ) -> Self

Synthesize a blank circuit.

fn from_frames( reduction_count: usize, frames: &[Self::EvalFrame], store: &'a Self::Store, folding_config: Arc<FoldingConfig<F, C>> ) -> Vec<Self>

Create an instance from some Self::Frames.

fn make_dummy( reduction_count: usize, circuit_frame: Option<Self::CircuitFrame>, store: &'a Self::Store, folding_config: Arc<FoldingConfig<F, C>>, meta: Meta<F> ) -> Self

Make a dummy instance, duplicating self’s final CircuitFrame.