Trait lurk::proof::Provable

pub trait Provable<F: LurkField> {
    // Required methods
    fn public_inputs(&self) -> Vec<F>;
    fn public_input_size() -> usize;
    fn reduction_count(&self) -> usize;
}

A trait for provable structures over a field F.

Required Methods

fn public_inputs(&self) -> Vec<F>

Returns the public inputs of the provable structure.

fn public_input_size() -> usize

Returns the size of the public inputs.

fn reduction_count(&self) -> usize

Returns the number of reductions in the provable structure.