Trait lurk::proof::Prover

pub trait Prover<'a, F: CurveCycleEquipped, C: Coprocessor<F> + 'a, M: MultiFrameTrait<'a, F, C>> {
    type PublicParams;
    type RecursiveSnark: RecursiveSNARKTrait<'a, F, C, M, PublicParams = Self::PublicParams>;

    // Required methods
    fn folding_mode(&self) -> &FoldingMode;
    fn reduction_count(&self) -> usize;
    fn lang(&self) -> &Arc<Lang<F, C>>;

    // Provided methods
    fn prove(
        &self,
        pp: &Self::PublicParams,
        frames: &[M::EvalFrame],
        store: &'a M::Store
    ) -> Result<(Self::RecursiveSnark, Vec<F>, Vec<F>, usize), ProofError> { ... }
    fn evaluate_and_prove(
        &self,
        pp: &Self::PublicParams,
        expr: M::Ptr,
        env: M::Ptr,
        store: &'a M::Store,
        limit: usize
    ) -> Result<(Self::RecursiveSnark, Vec<F>, Vec<F>, usize), ProofError> { ... }
    fn expected_num_steps(&self, raw_iterations: usize) -> usize { ... }
    fn outer_synthesize(
        &self,
        multiframes: &[M]
    ) -> Result<Vec<(M, TestConstraintSystem<F>)>, SynthesisError> { ... }
}

A trait for a prover that works with a field F.

Required Associated Types

type PublicParams

Associated type for public parameters

type RecursiveSnark: RecursiveSNARKTrait<'a, F, C, M, PublicParams = Self::PublicParams>

Assiciated proof type, which must implement RecursiveSNARKTrait

Required Methods

fn folding_mode(&self) -> &FoldingMode

Returns a reference to the prover’s FoldingMode

fn reduction_count(&self) -> usize

Returns the number of reductions for the prover.

fn lang(&self) -> &Arc<Lang<F, C>>

Returns a reference to the Prover’s Lang.

Provided Methods

fn prove( &self, pp: &Self::PublicParams, frames: &[M::EvalFrame], store: &'a M::Store ) -> Result<(Self::RecursiveSnark, Vec<F>, Vec<F>, usize), ProofError>

Generate a proof from a sequence of frames

fn evaluate_and_prove( &self, pp: &Self::PublicParams, expr: M::Ptr, env: M::Ptr, store: &'a M::Store, limit: usize ) -> Result<(Self::RecursiveSnark, Vec<F>, Vec<F>, usize), ProofError>

Evaluate an expression with an environment and then generate the corresponding proof

fn expected_num_steps(&self, raw_iterations: usize) -> usize

Returns the expected total number of steps for the prover given raw iterations.

fn outer_synthesize( &self, multiframes: &[M] ) -> Result<Vec<(M, TestConstraintSystem<F>)>, SynthesisError>

Synthesizes the outer circuit for the prover given a slice of multiframes.