Trait lurk::proof::Prover

pub trait Prover<'a, F: LurkField, C: Coprocessor<F> + 'a, M: MultiFrameTrait<'a, F, C>> {
    type PublicParams;

    // Required methods
    fn new(reduction_count: usize, lang: Lang<F, C>) -> Self;
    fn reduction_count(&self) -> usize;
    fn lang(&self) -> &Lang<F, C>;

    // Provided methods
    fn needs_frame_padding(&self, total_frames: usize) -> bool { ... }
    fn frame_padding_count(&self, total_frames: usize) -> usize { ... }
    fn expected_total_iterations(&self, raw_iterations: usize) -> usize { ... }
    fn multiframe_padding_count(&self, _raw_multiframe_count: usize) -> usize { ... }
    fn needs_multiframe_padding(&self, raw_multiframe_count: usize) -> bool { ... }
    fn outer_synthesize(
        &self,
        multiframes: &[M]
    ) -> Result<Vec<(M, TestConstraintSystem<F>)>, SynthesisError> { ... }
}

A trait for a prover that works with a field F.

Required Associated Types

type PublicParams

The associated public parameters type for the prover.

Required Methods

fn new(reduction_count: usize, lang: Lang<F, C>) -> Self

Creates a new prover with the specified number of reductions.

fn reduction_count(&self) -> usize

Returns the number of reductions for the prover.

fn lang(&self) -> &Lang<F, C>

Returns a reference to the Prover’s Lang.

Provided Methods

fn needs_frame_padding(&self, total_frames: usize) -> bool

Determines if the prover needs padding for a given total number of frames.

fn frame_padding_count(&self, total_frames: usize) -> usize

Returns the number of padding frames needed for a given total number of frames.

fn expected_total_iterations(&self, raw_iterations: usize) -> usize

Returns the expected total number of iterations for the prover given raw iterations.

fn multiframe_padding_count(&self, _raw_multiframe_count: usize) -> usize

Returns the number of padding multiframes needed for a given raw multiframe count.

fn needs_multiframe_padding(&self, raw_multiframe_count: usize) -> bool

Determines if the prover needs padding for a given raw multiframe count.

fn outer_synthesize( &self, multiframes: &[M] ) -> Result<Vec<(M, TestConstraintSystem<F>)>, SynthesisError>

Synthesizes the outer circuit for the prover given a slice of multiframes.

Object Safety

This trait is not object safe.