Type Alias lurk::ptr::Ptr

pub type Ptr<F> = GPtr<F, ExprTag>;

Aliased Type

struct Ptr<F> {
    pub tag: ExprTag,
    pub raw: RawPtr,
    pub _f: PhantomData<F>,
}

Fields
tag: ExprTag

An expression tag

raw: RawPtr

The underlying pointer, which can be null, opaque, or an index

_f: PhantomData<F>

PhantomData is needed to consume the F: LurkField parameter, since we want to pin our Ptr to a specific field (even though we don’t actually use it)

Implementations

impl<F: LurkField> Ptr<F>

pub fn allocate_maybe_fun_unconstrained<CS: ConstraintSystem<F>>( cs: CS, store: &Store<F>, maybe_fun: Option<&Ptr<F>> ) -> Result<(AllocatedNum<F>, AllocatedPtr<F>, AllocatedPtr<F>, AllocatedPtr<F>), SynthesisError>

impl<F: LurkField, T: Tag> GPtr<F, T>

pub const fn is_opaque(&self) -> bool

check if a Ptr is an opaque pointer

pub fn index(tag: T, idx: usize) -> Self

Construct a Ptr from an index

pub fn opaque(tag: T, idx: usize) -> Self

Construct a Ptr from an opaque index

pub fn null(tag: T) -> Self

Construct a null Ptr

impl<F: LurkField> GPtr<F, ExprTag>

pub fn cast(self, tag: ExprTag) -> Self

pub const fn as_cons(self) -> Option<Self>

pub const fn as_list(self) -> Option<Self>

pub const fn is_nil(&self) -> bool

check if a Ptr is Nil pointer

pub const fn is_cons(&self) -> bool

check if a Ptr is a Cons pointer

pub const fn is_atom(&self) -> bool

check if a Ptr is atomic pointer

pub const fn is_list(&self) -> bool