Struct lurk::store::Store

pub struct Store<F: LurkField> {

    pub cons_store: FrozenIndexSet<Box<(Ptr<F>, Ptr<F>)>, RandomState>,
    pub comm_store: FrozenIndexSet<Box<(FWrap<F>, Ptr<F>)>, RandomState>,
    pub fun_store: FrozenIndexSet<Box<(Ptr<F>, Ptr<F>, Ptr<F>)>, RandomState>,
    pub sym_store: FrozenIndexSet<Box<(Ptr<F>, Ptr<F>)>, RandomState>,
    pub num_store: FrozenIndexSet<Box<Num<F>>, RandomState>,
    pub str_store: FrozenIndexSet<Box<(Ptr<F>, Ptr<F>)>, RandomState>,
    pub thunk_store: FrozenIndexSet<Box<Thunk<F>>, RandomState>,
    pub call0_store: FrozenIndexSet<Box<(Ptr<F>, ContPtr<F>)>, RandomState>,
    pub call_store: FrozenIndexSet<Box<(Ptr<F>, Ptr<F>, ContPtr<F>)>, RandomState>,
    pub call2_store: FrozenIndexSet<Box<(Ptr<F>, Ptr<F>, ContPtr<F>)>, RandomState>,
    pub tail_store: FrozenIndexSet<Box<(Ptr<F>, ContPtr<F>)>, RandomState>,
    pub lookup_store: FrozenIndexSet<Box<(Ptr<F>, ContPtr<F>)>, RandomState>,
    pub unop_store: FrozenIndexSet<Box<(Op1, ContPtr<F>)>, RandomState>,
    pub binop_store: FrozenIndexSet<Box<(Op2, Ptr<F>, Ptr<F>, ContPtr<F>)>, RandomState>,
    pub binop2_store: FrozenIndexSet<Box<(Op2, Ptr<F>, ContPtr<F>)>, RandomState>,
    pub if_store: FrozenIndexSet<Box<(Ptr<F>, ContPtr<F>)>, RandomState>,
    pub let_store: FrozenIndexSet<Box<(Ptr<F>, Ptr<F>, Ptr<F>, ContPtr<F>)>, RandomState>,
    pub letrec_store: FrozenIndexSet<Box<(Ptr<F>, Ptr<F>, Ptr<F>, ContPtr<F>)>, RandomState>,
    pub emit_store: FrozenIndexSet<Box<ContPtr<F>>, RandomState>,
    pub opaque_ptrs: FrozenIndexSet<Box<ZExprPtr<F>>, RandomState>,
    pub opaque_cont_ptrs: FrozenIndexSet<Box<ZContPtr<F>>, RandomState>,
    pub z_expr_ptr_map: FrozenMap<ZExprPtr<F>, Box<Ptr<F>>>,
    pub z_cont_ptr_map: FrozenMap<ZContPtr<F>, Box<ContPtr<F>>>,
    pub poseidon_cache: PoseidonCache<F>,
    pub inverse_poseidon_cache: InversePoseidonCache<F>,
    pub dehydrated: ArcSwap<FrozenVec<Box<Ptr<F>>>>,
    pub dehydrated_cont: ArcSwap<FrozenVec<Box<ContPtr<F>>>>,
    pub constants: OnceCell<NamedConstants<F>>,
    /* private fields */

}

Fields
cons_store: FrozenIndexSet<Box<(Ptr<F>, Ptr<F>)>, RandomState>comm_store: FrozenIndexSet<Box<(FWrap<F>, Ptr<F>)>, RandomState>fun_store: FrozenIndexSet<Box<(Ptr<F>, Ptr<F>, Ptr<F>)>, RandomState>sym_store: FrozenIndexSet<Box<(Ptr<F>, Ptr<F>)>, RandomState>

Holds a Sym or Key which is a string head and a symbol tail

num_store: FrozenIndexSet<Box<Num<F>>, RandomState>str_store: FrozenIndexSet<Box<(Ptr<F>, Ptr<F>)>, RandomState>

Holds a Str, which is a char head and a string tail

thunk_store: FrozenIndexSet<Box<Thunk<F>>, RandomState>call0_store: FrozenIndexSet<Box<(Ptr<F>, ContPtr<F>)>, RandomState>call_store: FrozenIndexSet<Box<(Ptr<F>, Ptr<F>, ContPtr<F>)>, RandomState>call2_store: FrozenIndexSet<Box<(Ptr<F>, Ptr<F>, ContPtr<F>)>, RandomState>tail_store: FrozenIndexSet<Box<(Ptr<F>, ContPtr<F>)>, RandomState>lookup_store: FrozenIndexSet<Box<(Ptr<F>, ContPtr<F>)>, RandomState>unop_store: FrozenIndexSet<Box<(Op1, ContPtr<F>)>, RandomState>binop_store: FrozenIndexSet<Box<(Op2, Ptr<F>, Ptr<F>, ContPtr<F>)>, RandomState>binop2_store: FrozenIndexSet<Box<(Op2, Ptr<F>, ContPtr<F>)>, RandomState>if_store: FrozenIndexSet<Box<(Ptr<F>, ContPtr<F>)>, RandomState>let_store: FrozenIndexSet<Box<(Ptr<F>, Ptr<F>, Ptr<F>, ContPtr<F>)>, RandomState>letrec_store: FrozenIndexSet<Box<(Ptr<F>, Ptr<F>, Ptr<F>, ContPtr<F>)>, RandomState>emit_store: FrozenIndexSet<Box<ContPtr<F>>, RandomState>opaque_ptrs: FrozenIndexSet<Box<ZExprPtr<F>>, RandomState>

Holds opaque pointers

opaque_cont_ptrs: FrozenIndexSet<Box<ZContPtr<F>>, RandomState>

Holds opaque continuation pointers

z_expr_ptr_map: FrozenMap<ZExprPtr<F>, Box<Ptr<F>>>

Holds a mapping of ZExprPtr -> Ptr for reverse lookups

z_cont_ptr_map: FrozenMap<ZContPtr<F>, Box<ContPtr<F>>>

Holds a mapping of ZExprPtr -> ContPtr<F> for reverse lookups

poseidon_cache: PoseidonCache<F>

Caches poseidon hashes

inverse_poseidon_cache: InversePoseidonCache<F>

Caches poseidon preimages

dehydrated: ArcSwap<FrozenVec<Box<Ptr<F>>>>

Contains Ptrs which have not yet been hydrated.

dehydrated_cont: ArcSwap<FrozenVec<Box<ContPtr<F>>>>constants: OnceCell<NamedConstants<F>>
Implementations

impl<F: LurkField> Store<F>

pub fn as_lurk_boolean(&self, x: bool) -> Ptr<F>

impl<F: LurkField> Store<F>

pub fn read(&self, input: &str) -> Result<Ptr<F>, Error>

pub fn read_with_state( &self, state: Rc<RefCell<State>>, input: &str ) -> Result<Ptr<F>, Error>

pub fn read_maybe_meta_with_state<'a>( &self, state: Rc<RefCell<State>>, input: Span<'a> ) -> Result<(Span<'a>, Ptr<F>, bool), Error>

impl<F: LurkField> Store<F>

These methods provide a more ergonomic means of constructing and manipulating Lurk data. They can be thought of as a minimal DSL for working with Lurk data in Rust code. Prefer these methods when constructing literal data or assembling program fragments in tests or during evaluation, etc.

pub fn expect_constants(&self) -> &NamedConstants<F>

pub fn cons(&self, car: Ptr<F>, cdr: Ptr<F>) -> Ptr<F>

pub fn strcons(&self, car: Ptr<F>, cdr: Ptr<F>) -> Ptr<F>

pub fn strnil(&self) -> Ptr<F>

pub fn symnil(&self) -> Ptr<F>

pub fn hidden(&self, secret: F, payload: Ptr<F>) -> Option<Ptr<F>>

pub fn hide(&self, secret: F, payload: Ptr<F>) -> Ptr<F>

pub fn commit(&self, payload: Ptr<F>) -> Ptr<F>

pub fn open(&self, ptr: Ptr<F>) -> Option<(F, Ptr<F>)>

pub fn open_mut(&self, ptr: Ptr<F>) -> Result<(F, Ptr<F>), Error>

pub fn secret(&self, ptr: Ptr<F>) -> Option<Ptr<F>>

pub fn secret_mut(&self, ptr: Ptr<F>) -> Result<Ptr<F>, Error>

pub fn list(&self, elts: &[Ptr<F>]) -> Ptr<F>

pub fn num<T: Into<Num<F>>>(&self, num: T) -> Ptr<F>

pub fn uint64(&self, n: u64) -> Ptr<F>

pub fn str(&self, s: &str) -> Ptr<F>

pub fn sym<T: AsRef<str>>(&self, name: T) -> Ptr<F>

pub fn key<T: AsRef<str>>(&self, name: T) -> Ptr<F>

pub fn car(&self, expr: &Ptr<F>) -> Result<Ptr<F>, Error>

pub fn cdr(&self, expr: &Ptr<F>) -> Result<Ptr<F>, Error>

pub fn new() -> Self

pub fn intern_symnil(&self, key: bool) -> Ptr<F>

pub fn intern_cons(&self, car: Ptr<F>, cdr: Ptr<F>) -> Ptr<F>

pub fn intern_strcons(&self, car: Ptr<F>, cdr: Ptr<F>) -> Ptr<F>

pub fn intern_symcons(&self, car: Ptr<F>, cdr: Ptr<F>) -> Ptr<F>

pub fn intern_keycons(&self, car: Ptr<F>, cdr: Ptr<F>) -> Ptr<F>

pub fn intern_comm(&self, secret: F, payload: Ptr<F>) -> Ptr<F>

pub fn intern_maybe_opaque(&self, tag: ExprTag, hash: F) -> Ptr<F>

pub fn get_maybe_opaque(&self, tag: ExprTag, hash: F) -> Option<Ptr<F>>

pub fn intern_maybe_opaque_fun(&self, hash: F) -> Ptr<F>

pub fn intern_maybe_opaque_sym(&self, hash: F) -> Ptr<F>

pub fn intern_maybe_opaque_cons(&self, hash: F) -> Ptr<F>

pub fn intern_maybe_opaque_comm(&self, hash: F) -> Ptr<F>

pub fn intern_opaque_fun(&self, hash: F) -> Ptr<F>

pub fn intern_opaque_sym(&self, hash: F) -> Ptr<F>

pub fn intern_opaque_cons(&self, hash: F) -> Ptr<F>

pub fn intern_opaque_comm(&self, hash: F) -> Ptr<F>

pub fn intern_list(&self, elts: &[Ptr<F>]) -> Ptr<F>

Helper to allocate a list, instead of manually using cons.

pub fn intern_symbol_path(&self, path: &[String]) -> Ptr<F>

pub fn intern_symbol(&self, sym: &Symbol) -> Ptr<F>

pub fn user_sym(&self, name: &str) -> Ptr<F>

pub fn intern_num<T: Into<Num<F>>>(&self, num: T) -> Ptr<F>

pub fn get_num<T: Into<Num<F>>>(&self, num: T) -> Option<Ptr<F>>

pub fn intern_char(&self, c: char) -> Ptr<F>

pub fn intern_uint(&self, n: UInt) -> Ptr<F>

pub fn intern_u64(&self, n: u64) -> Ptr<F>

pub fn intern_string(&self, s: &str) -> Ptr<F>

pub fn intern_fun( &self, arg: Ptr<F>, body: Ptr<F>, closed_env: Ptr<F> ) -> Ptr<F>

pub fn intern_thunk(&self, thunk: Thunk<F>) -> Ptr<F>

pub fn mark_dehydrated_cont(&self, p: ContPtr<F>) -> ContPtr<F>

pub fn get_cont_outermost(&self) -> ContPtr<F>

pub fn get_cont_error(&self) -> ContPtr<F>

pub fn get_cont_terminal(&self) -> ContPtr<F>

pub fn get_cont_dummy(&self) -> ContPtr<F>

pub fn get_hash_components_cont(&self, ptr: &ContPtr<F>) -> Option<[F; 8]>

pub fn get_hash_components_thunk(&self, thunk: &Thunk<F>) -> Option<[F; 4]>

pub fn intern_cont_error(&self) -> ContPtr<F>

pub fn intern_cont_outermost(&self) -> ContPtr<F>

pub fn intern_cont_terminal(&self) -> ContPtr<F>

pub fn intern_cont_dummy(&self) -> ContPtr<F>

pub fn fetch_z_expr_ptr(&self, z_ptr: &ZExprPtr<F>) -> Option<Ptr<F>>

pub fn fetch_z_cont_ptr(&self, z_ptr: &ZContPtr<F>) -> Option<ContPtr<F>>

pub fn fetch_maybe_sym(&self, ptr: &Ptr<F>) -> Option<Symbol>

pub fn fetch_symcons(&self, ptr: &Ptr<F>) -> Option<(Ptr<F>, Ptr<F>)>

pub fn fetch_sym(&self, ptr: &Ptr<F>) -> Option<Symbol>

pub fn fetch_key(&self, ptr: &Ptr<F>) -> Option<Symbol>

pub fn fetch_symbol(&self, ptr: &Ptr<F>) -> Option<Symbol>

pub fn fetch_strcons(&self, ptr: &Ptr<F>) -> Option<(Ptr<F>, Ptr<F>)>

pub fn fetch_string(&self, ptr: &Ptr<F>) -> Option<String>

pub fn fetch_char(&self, ptr: &Ptr<F>) -> Option<char>

pub fn fetch_fun(&self, ptr: &Ptr<F>) -> Option<&(Ptr<F>, Ptr<F>, Ptr<F>)>

pub fn fetch_cons(&self, ptr: &Ptr<F>) -> Option<&(Ptr<F>, Ptr<F>)>

pub fn fetch_comm(&self, ptr: &Ptr<F>) -> Option<&(FWrap<F>, Ptr<F>)>

pub fn fetch_num(&self, ptr: &Ptr<F>) -> Option<&Num<F>>

pub fn fetch_thunk(&self, ptr: &Ptr<F>) -> Option<&Thunk<F>>

pub fn fetch_uint(&self, ptr: &Ptr<F>) -> Option<UInt>

pub fn fetch(&self, ptr: &Ptr<F>) -> Option<Expression<F>>

pub fn fetch_list(&self, ptr: &Ptr<F>) -> Option<Vec<Ptr<F>>>

Returns a Vec of Ptrs representing the elements of a proper list, ptr. This is intended to be the inverse of Store::list(). IF ptr isn’t a proper list, return None.

pub fn fetch_cont(&self, ptr: &ContPtr<F>) -> Option<Continuation<F>>

pub fn get_z_expr( &self, ptr: &Ptr<F>, z_store: &mut Option<ZStore<F>> ) -> Result<(ZExprPtr<F>, Option<ZExpr<F>>), Error>

pub fn get_z_cont( &self, ptr: &ContPtr<F>, z_store: &mut Option<ZStore<F>> ) -> Result<(ZContPtr<F>, Option<ZCont<F>>), Error>

pub fn to_z_store_with_ptr( &self, ptr: &Ptr<F> ) -> Result<(ZStore<F>, ZExprPtr<F>), Error>

pub fn to_z_expr(&self, ptr: &Ptr<F>) -> Option<ZExpr<F>>

pub fn hash_expr(&self, ptr: &Ptr<F>) -> Option<ZExprPtr<F>>

pub fn to_z_cont(&self, ptr: &ContPtr<F>) -> Option<ZCont<F>>

pub fn hash_cont(&self, ptr: &ContPtr<F>) -> Option<ZContPtr<F>>

pub fn hash_string(&self, s: &str) -> ZExprPtr<F>

pub fn hash_symbol(&self, s: &Symbol) -> ZExprPtr<F>

pub fn car_cdr(&self, ptr: &Ptr<F>) -> Result<(Ptr<F>, Ptr<F>), Error>

pub fn get_opaque_ptr(&self, ptr: Ptr<F>) -> Option<ZExprPtr<F>>

pub fn new_opaque_ptr(&self) -> Ptr<F>

pub fn ptr_eq(&self, a: &Ptr<F>, b: &Ptr<F>) -> Result<bool, Error>

pub fn cons_eq(&self, a: &Ptr<F>, b: &Ptr<F>) -> bool

pub fn hydrate_scalar_cache(&self)

Fill the cache for Scalars. Only Ptrs which have been interned since last hydration will be hashed, so it is safe to call this incrementally. However, for best proving performance, we should call exactly once so all hashing can be batched, e.g. on the GPU.

pub fn z_expr_ptr_from_parts( &self, tag: F, value: F ) -> Result<ZExprPtr<F>, Error>

pub fn z_cont_ptr_from_parts( &self, tag: F, value: F ) -> Result<ZContPtr<F>, Error>

pub fn intern_z_expr_ptr( &self, z_ptr: &ZExprPtr<F>, z_store: &ZStore<F> ) -> Option<Ptr<F>>

pub fn intern_z_cont_ptr( &self, z_ptr: &ZContPtr<F>, z_store: &ZStore<F> ) -> Option<ContPtr<F>>

impl<F: LurkField> Store<F>

pub fn intern_syntax(&self, syn: Syntax<F>) -> Ptr<F>