Trait lurk::tag::Tag

pub trait Tag: Into<u16> + TryFrom<u16, Error = Error> + Copy + Sized + Hash + Eq + Debug {
    // Required methods
    fn from_field<F: LurkField>(f: &F) -> Option<Self>;
    fn to_field<F: LurkField>(&self) -> F;

    // Provided method
    fn to_field_bytes<F: LurkField>(&self) -> F::Repr { ... }
}

Required Methods

fn from_field<F: LurkField>(f: &F) -> Option<Self>

fn to_field<F: LurkField>(&self) -> F

Provided Methods

fn to_field_bytes<F: LurkField>(&self) -> F::Repr

This explicitly defines a shortcut to access the canonical byte representation of the field element associated to a tag. We expect some tag types to be able to improve upon it.